==> scripts/01_clean_pheno.Rmd <==
---
title: "I mean really clean the pheno and get to know the hell of the samples"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, message = F, warning = F)
library(tidyverse)
library(readxl)
library(RVenn)
```
 
## Valid patient samples

```{r}
# the newest pheno table from M
# this sheet have nice sampleids for the samples, and also this is the table has all of the info
phe_M <- read_excel('/Volumes/castoricenter/CAR-Tcell_and_Microbiota/CTC Event Table 10232019 wo Censor .xlsx',
                    sheet = 5)
```

```{r}
# selecting the columns relavant to this study
phe_M_clean <- phe_M %>% 
   rename_all(funs(str_replace_all(., ' \\(.+$','') %>% 
    str_replace_all(., ' ','_') %>% 
    str_replace(., ':','_'))) %>% 
  # select the columns that may be relavant
  select(Sampleid = Baseline_Sample_ID, 
         Censor_reason,
         CR, 
         ORR, 
         Toxicity, 
         Disease_Status_CAR = Key__Patient_Disease_Status_at_CAR, 
         Disease_Class, 
         CAR_Group) %>% 
  filter(is.na(Censor_reason)) 

nrow(phe_M_clean)

head(phe_M_clean)


# output this pheno table cuz its gonna be used downstream
# this is all of the baseline samples that has 16s.
phe_M_clean %>% 
  write_csv('../data/cleaned_pheno/current_pheno.csv')

pheno <- read_csv('../data/cleaned_pheno/final_pheno.csv')
table(pheno$CR)
```


## Valid shotgun data


```{r}
# loading the table from John to see how many samples we actually have shotgun data
have_sg <- read_csv('../data/baselinesforAngel10032019.csv') %>% 
  filter(shotgun_YN == 'yes') %>% 
  # removig 1943A cuz its low quality
  filter(sample_id != '1943A') %>% 
  ### FUCK THIS TABLE OF ../data/baselinesforAngel10032019.csv IS WRONG IN 161, it didn’t pass QC because not enough DNA material was present. SO WE DON'T HAVE THE SHOTGUN DATA.
  filter(!sample_id %in% c('161'))

nrow(have_sg)
  


```

## Overlap: the real valid shotgun samples

```{r}
# visualize the difference 
two <- Venn(list(
  have_file = have_sg$sample_id,
  useable = phe_M_clean$Sampleid
))  
  
ggvenn(two)      


# so the final shotgun samples number will be 33
length(intersect(have_sg$sample_id, phe_M_clean$Sampleid))

final_sample_id <- intersect(have_sg$sample_id, phe_M_clean$Sampleid)

# so the final removed 2 samples are:
setdiff(pheno_col$Sampleid, final_sample_id) 
```

```{r}
# the final pheno table which has 33 rows
final_pheno <- phe_M_clean %>% 
  filter(Sampleid %in% have_sg$sample_id)

final_pheno %>% 
  write_csv('../data/cleaned_pheno/final_pheno.csv') 

# write out the meta data info with tox and CR in each of the table
final_pheno %>% 
  select(Sampleid, Toxicity) %>% 
  write_csv('../data/cleaned_pheno/melody_tox_metadata.csv')
```


```{r}
# write out the overlap sampleid cuz that's gonna be the sample_names in use on the cluster for normalization
intersect(have_sg$sample_id, phe_M_clean$Sampleid) %>%  
  write.table('../data/input_for_cluster_sampleid/valid_sample_names.txt', col.names = F, row.names = F, quote = F)
```


```{r}
# retrieve the toxicity grade info from the event table
events <- read_excel('/Volumes/castoricenter/CAR-Tcell_and_Microbiota/CTC Event Table 10232019 wo Censor .xlsx',
                    sheet = 3) %>% 
  select(MRN, Event, `Event Description`, `Event Stop`) %>% 
  rename(descr = names(.)[3],
         stop = names(.)[4]) %>% 
  filter(Event %in% c('CRS','ICANS')) %>% 
  mutate(grade = if_else(str_detect(descr, '^No'), '0', str_extract(descr, '[1|2|3]'))) %>% 
  select(-descr)

table(events$Event)
  
# 35402481 patient ICANS record missing.

# try with the CRS first
tox_CRS <- events %>% 
  filter(Event == 'CRS')

table(tox_CRS$grade)

# try to get the sample id from MRN
# this part was manually done by me cleaning M's table

# below is for CRS only
samps <- read_csv('../data/cleaned_pheno/MRN_sampleid.txt', col_types = 'cc') %>% 
  left_join(tox_CRS, by  = 'MRN') %>% 
  filter(sampleid %in% pheno$Sampleid) %>% 
  # really mad that i have to remove some records in this way
  mutate(stop = replace(stop, MRN == '35546332' & grade == '0', '2018-01-30')) %>% 
  arrange(desc(stop)) %>% 
  distinct(sampleid, .keep_all = T) %>% 
  rename(CRS_grade = grade) %>% 
  select(-stop)

# 35546332 has both None and grade 1 for CRS
# based on the date choose the 0
# choose the latest date grade for 00332772 : 1
# choose the latest date grade for 35546332 : 0  

samps %>% 
  write_csv('../data/cleaned_pheno/cleaned_CRS.csv')
```
