==> scripts/05_pathway_feature_cpm.Rmd <==
---
title: "Check the pathway cpm of the feature"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(tidyverse)
library(ggpubr)
```

Check the pathway abundance in cpm to see if there is a decent effect size

```{r}
pwa <- read_tsv('../data/humann2_res/humann2_final_out/humann2_pathabundance_cpm_unstratified.tsv') %>%
  dplyr::rename(pw = names(.)[1]) %>% 
  rename_all(funs(str_replace(., '_Abundance$',''))) %>% 
  # create a new column just to be pwid 
  mutate(pwid = str_extract(pw, '^.+:')) %>% 
  mutate(pwid = str_replace(pwid, ':',''))

  # removing things in the parenthesis and remove the spaces in the pathway
  # mutate(pathway = str_replace_all(pathway, ' \\(.+\\)',''),
  #        pathway = str_replace_all(pathway, ' ',''))

pheno <- read_csv('../data/cleaned_pheno/final_pheno.csv') %>% 
  dplyr::select(sampleid = Sampleid, 
         Toxicity,
         CR)
```
 
```{r}
sig <- read_csv('../data/unique_sig_pathway.csv') %>% 
  # extract the pw id part
  mutate(pwid = str_extract(pathway, '^.+_PWY|^PWY.+\\d{3,4}'))  %>% 
  mutate(pwid = str_replace_all(pwid, '_', '-')) 


pwa_sig <- pwa %>% 
  filter(pwid %in% sig$pwid) %>% 
  dplyr::select(-pwid) %>% 
  gather(key = 'sampleid', value = 'abun', names(.)[2]:names(.)[ncol(.)]) %>% 
  left_join(pheno, by  = 'sampleid')
```


```{r}
# bar plot to show the cpm in each sample for each sig pathway
pw_list <- pwa_sig %>% 
  distinct(pw) %>% 
  pull(pw)

# tox
for (p in pw_list) {
  pwa_sig %>% 
    arrange(Toxicity, desc(abun)) %>% 
        filter(pw == p) %>% 
        ggbarplot(x = 'sampleid', y = 'abun', fill = 'Toxicity', color = 'Toxicity') +
          scale_color_manual(values = c('#00468B', '#EC0000'))  +
          scale_fill_manual(values = c('#00468B', '#EC0000')) +
          theme(axis.text.x = element_text(angle = 45, hjust = 1, size = 8)) +
          labs(title = str_glue('Pathway: {p}')) +
          ggsave(str_glue('/Volumes/castoricenter/CAR-Tcell_and_Microbiota/humann2_shotgun/figs/pathway_abundance_cpm/Toxicity/pathway_abundance_cpm_{p}.jpg'), width = 12, height = 10, dpi = 300)
}
```


```{r}
# CR
for (p in pw_list) {
  pwa_sig %>% 
    arrange(CR, desc(abun)) %>% 
        filter(pw == p) %>% 
        ggbarplot(x = 'sampleid', y = 'abun', fill = 'CR', color = 'CR') +
          scale_color_manual(values = c('#00468B', '#EC0000'))  +
          scale_fill_manual(values = c('#00468B', '#EC0000')) +
          theme(axis.text.x = element_text(angle = 45, hjust = 1, size = 8)) +
          labs(title = str_glue('Pathway: {p}')) +
          ggsave(str_glue('/Volumes/castoricenter/CAR-Tcell_and_Microbiota/humann2_shotgun/figs/pathway_abundance_cpm/CR/pathway_abundance_cpm_{p}.jpg'), width = 12, height = 10, dpi = 300)
}
```
