==> scripts/06_lefse_res_ggplot.Rmd <==
---
title: "Replot the lefse plots with ggplot"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(tidyverse)
library(grid)
library(cowplot)
library(ggpubr)
```

Replot the lefse plots with ggplot2 (lefse with only class )

## Combine and clean the res table from lefse

```{r}
# the input : lefse res files
fns <- list.files('../data/lefse_input/', pattern = 'pathway_.+res$', full.names = T)

fns_name <- list.files('../data/lefse_input/', pattern = 'pathway_.+res$')

# join all of the tables feature together
feature <- fns %>% 
  set_names(fns_name) %>% 
  map(~ read_tsv(., col_names = c('pathway','xx','direction','score','pval')) %>% 
        filter(!is.na(score))) %>% 
  bind_rows(.id = 'group') %>% 
  mutate(group = str_replace(group, 'pathway_abundace_pheno__','')) %>% 
  mutate(group = str_replace(group, '.res$','')) 

# change the "N" direction to be minus score
feature <- bind_rows(
  feature %>% 
    split(.$direction) %>% 
    pluck('N') %>% 
    mutate(score = -score),
  feature %>% 
    split(.$direction) %>% 
    pluck('Y')  
) %>% 
  arrange(group, pathway, score)

feature %>% 
  write_csv('../data/lefse_results_sig_pathway.csv')
```

```{r}
# output the sig unique pathway from the above table for later use in the bar plot
feature  %>% 
  distinct(pathway, .keep_all = T) %>% 
  write_csv('../data/unique_sig_pathway.csv')
```


## Recreate the lefse figure with ggplot

Each combi per figure


```{r}
all_title_fs <- 20
axis_text_fs <- 16

CR <- feature %>% 
  filter(group == 'CR') %>% 
  ggplot(aes(x = reorder(pathway, score), y = score, fill = direction)) +
            geom_bar( stat = 'identity') +
            coord_flip() +
            scale_color_manual(values = c('#925E9F', '#42B540'))  +
            scale_fill_manual(values = c('#925E9F', '#42B540')) +
            theme_classic() +
            theme(axis.title.y  = element_blank(), 
                  plot.title = element_text(size=all_title_fs),
                  axis.title.x = element_text(size=axis_text_fs),
                  axis.text.x = element_text(size=axis_text_fs),
                  legend.position='bottom') +
            labs(title = str_glue('Response to CR') ,
                 y = 'Score')+ 
            ggsave(str_glue('/Volumes/castoricenter/CAR-Tcell_and_Microbiota/humann2_shotgun/figs/lefse_results_ggplot_recreate/Response to CR.pdf'),  width = 12, device = 'pdf' )

tox <- feature %>% 
  filter(group == 'Toxicity') %>% 
  ggplot(aes(x = reorder(pathway, score), y = score, fill = direction)) +
            geom_bar(stat = 'identity') +
            coord_flip() +
            scale_color_manual(values = c('#0099B4', '#AD002A'))  +
            scale_fill_manual(values = c('#0099B4', '#AD002A')) +
            theme_classic() +
            theme(axis.title.y  = element_blank(), 
                  axis.title.x = element_text(size=axis_text_fs),
                  plot.title = element_text(size=all_title_fs),
                  axis.text.x = element_text(size=axis_text_fs),
                  legend.position="bottom") +
            labs(title = str_glue('Response to Toxicity') ,
                 y = 'Score')+ 
            ggsave(str_glue('/Volumes/castoricenter/CAR-Tcell_and_Microbiota/humann2_shotgun/figs/lefse_results_ggplot_recreate/Response to tox.pdf'), device = 'pdf', width = 12)


```


```{r}
# stack the above two plots vertically
g <- plot_grid(CR,tox,  
          nrow = 2, 
          align = 'hv',
          axis = 'b')  +
  ggsave('/Volumes/castoricenter/CAR-Tcell_and_Microbiota/humann2_shotgun/figs/lefse_results_ggplot_recreate/pathway_lefse_combined.pdf', device = 'pdf', height = 15, width = 15)
```
