==> scripts/08_barplot_species_pathway.Rmd <==
---
title: "Bar plot species contribution to pathway with humann2 script"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(tidyverse)
library(RColorBrewer)
```

humann2_barplot
Basic usage: *humann2_barplot --input TABLE.tsv --feature FEATURE --outfile $FIGURE*
$TABLE.tsv = a stratified HUMAnN2 output file
$FEATURE = Feature from the table to plot (defaults to first feature)
$FIGURE = Where to save the figure
Run with -h to see additional command line options
humann2_barplot produces plots of stratified HUMAnN2 features. Note: unlike many other HUMAnN2 utilities, humann2_barplot requires the Python scientific stack (notably matplotlib) to operate. humann2_barplot includes many options for sorting and scaling data, as detailed in the --help menu. humann2_barplot is also explored in greater detail in the HUMAnN2 Tutorial.

Here is an example of a HUMAnN2 barplot for a pathway (denitrification) that was preferentially enriched in Human Microbiome Project oral samples relative to other body sites. This figure uses many options from humann2_barplot, including *regrouping by genus*, *pseudolog scaling*, and *sorting samples by similarity and metadata*:


## Make species level bar plot using humann2 script

### Create the correct pheno and relab table format

```{r}
# ok so the correct format should be header at the first row, and phenotype and then the abundance values 
valid_s <- read_table('../data/input_for_cluster_sampleid/valid_sample_names.txt', col_names = F)
# pheno table 
read_csv('../data/cleaned_pheno/current_pheno.csv') %>% 
  filter(Sampleid %in% valid_s$X1) %>% 
  select(-Censor_YN, -Censor_reason) %>% 
  # USE THE Toxicity PHENO FIRST
  select(Sampleid, Toxicity) %>% 
  t() %>% 
  write.table('../data/horizontal_pheno/tox_wi_sample_header.txt', quote = F, row.names = T, col.names = F, sep = '\t')

# reordered relab table but without header
read_tsv('../data/humann2_res/humann2_final_out/humann2_pathabundance_relab_stratified_reordered.tsv') %>% 
  write_tsv('../data/humann2_res/humann2_final_out/humann2_pathabundance_relab_stratified_reordered_wo_header.tsv', col_names = F)
```

### row bind them together

```{bash}
# guess we need to cat the pheno and cnts together?
cat ../data/horizontal_pheno/tox_wi_sample_header.txt \
    ../data/humann2_res/humann2_final_out/humann2_pathabundance_relab_stratified_reordered_wo_header.tsv > \
    ../data/humann2_res/humann2_final_out/humann2_pathabundance_relab_stratified_reordered_with_tox.tsv
```

FINALLY WORKED!!!!!!!!!!!!!!!!!!!!!!

```{bash}
# the actual cmd to plot 
humann2_barplot  \
  --input ../data/humann2_res/humann2_final_out/humann2_pathabundance_relab_stratified_reordered_with_tox.tsv \
  --focal-feature PWY-4242 \
  --sort similarity sum metadata \
  --focal-metadatum Toxicity --last-metadatum Toxicity \
  --scaling pseudolog \
  --dimensions 10 5 \
  -o ../figs/barplot_humann2/PWY-4242_pw.png
```

### now make the cmd with str_glue

```{r}
# highlight the bar plots of tox in pathways only come out tox lefse
tox_pathway <- read_csv('../data/lefse_stats_all_sig_pathway.csv') %>% 
  filter(str_detect(group, '^Toxicity')) %>% 
  distinct(pathway, .keep_all = T) %>% 
  select(pathway)  %>% 
  mutate(pathway = str_replace(pathway, '^f_','')) %>% 
  mutate(pwid = str_extract(pathway, '^.+_PWY|^PWY_\\d{3,4}'))  %>% 
  mutate(pwid = str_replace_all(pwid, '_', '-')) 
  
```


# --exclude-unclassified

```{r}
# load the sig pathway table with pwid
cmd_barplot <- tox_pathway %>% 
  select(pwid) %>% 
  mutate(barcmd = str_glue('humann2_barplot  --input ../data/humann2_res/humann2_final_out/humann2_pathabundance_relab_stratified_reordered_with_tox.tsv --focal-feature {pwid} --sort similarity sum metadata --exclude-unclassified --focal-metadatum Toxicity --last-metadatum Toxicity --scaling pseudolog --dimensions 10 5 -o /Volumes/castoricenter/CAR-Tcell_and_Microbiota/humann2_shotgun/figs/species_contribution_barplot_humann2/without_unclassfified/Toxicity/{pwid}_pw_no_unclassfified.png')) %>% 
  select(barcmd) 
 
cmd_barplot %>% 
  write.table('humann2_bar_cmd_no_unclassfified_tox.sh', quote = F, col.names = F, row.names = F )
```




# has unclassified


```{r}
# with unclassfified
cmd_barplot <- tox_pathway %>% 
  select(pwid) %>% 
  mutate(barcmd = str_glue('humann2_barplot  --input ../data/humann2_res/humann2_final_out/humann2_pathabundance_relab_stratified_reordered_with_tox.tsv --focal-feature {pwid} --sort similarity sum metadata  --focal-metadatum Toxicity --last-metadatum Toxicity --scaling pseudolog --dimensions 10 5 -o /Volumes/castoricenter/CAR-Tcell_and_Microbiota/humann2_shotgun/figs/species_contribution_barplot_humann2/with_unclassfified/Toxicity/{pwid}_pw.png')) %>% 
  select(barcmd) 
 
cmd_barplot %>% 
  write.table('humann2_bar_cmd_tox.sh', quote = F, col.names = F, row.names = F )
```


## Make genus level bar plot 

Genus level gene families and pathways
By default, the gene families and pathways output files from HUMAnN2 are species level. To obtain genus level gene families and pathways, follow these steps.

Create a genus level gene families file

$ humann2_gene_families_genus_level --input $SAMPLE_genefamilies.tsv --output $SAMPLE_genefamilies_genus_level.tsv
In this command, replace $SAMPLE_genefamilies.tsv with the species level gene families file created by default by HUMAnN2 and $SAMPLE_genefamilies_genus_level.tsv with the name of the gene families genus level file that will be created.
Run HUMAnN2, with the genus level gene families file as input, to get genus level pathways output files

$ humann2 --input $SAMPLE_genefamilies_genus_level.tsv --output humann2_genus_level_output
This run will be much faster and require less memory than the original run as HUMAnN2 is provided gene family abundances so it only needs to compute the pathways.
