==> scripts/10_genefamilies_KO_regroup.Rmd <==
---
title: "Humann2 bar plot for gene family"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(tidyverse)
library(ggpubr)
```

######################################## Bash ########################################

```{bash}
# needs to download the utility mapping db first
humann2_databases --download utility_mapping full /Users/daia1/melodyP/data/humann2_res
```

```{bash}
# regroup to KEGG Orthogroups (KOs)
humann2_regroup_table \
            --input ../data/humann2_res/humann2_genefamilies.tsv  \
            --groups uniref90_ko  \
            --output ../data/humann2_res/humann2_genefamilies_KO_regroup.tsv
```

```{bash}
# renorm to relab
humann2_renorm_table  --input ../data/humann2_res/humann2_genefamilies_KO_regroup.tsv -u relab --output ../data/humann2_res/humann2_genefamilies_KO_regroup_relab.tsv
```

```{bash}
# split into unstratified and ..
humann2_split_stratified_table --input ../data/humann2_res/humann2_genefamilies_KO_regroup_relab.tsv  --output ../data/humann2_res/genefamily_split_KO
```

######################################## R ############################################

```{r} 

# reorder the unstratified table
genef_KO <- read_tsv('../data/humann2_res/genefamily_split_KO/humann2_genefamilies_KO_regroup_relab_unstratified.tsv')%>% 
  rename(KO = names(.)[1]) %>% 
  rename_all(funs(str_replace(., '_Abundance-RPKs$',''))) 
  

pheno <- read_csv('../data/current_pheno.csv') %>% 
  filter(! Sampleid %in% c("1725E", "2222A" ,"2245A"))

genef_KO <- bind_cols(genef_KO[,1],genef_KO[,pheno$Sampleid]) 
all.equal(pheno$Sampleid, colnames(genef_KO)[2:ncol(genef_KO)])
  
```
```{r}
patha
```


```{r}
# look at the relab abundance of the pathways related to pantothenate for example
patha <- read_tsv('../data/humann2_res/humann2_final_out/humann2_pathabundance_relab_unstratified.tsv') %>% 
  rename(pw = names(.)[1]) %>% 
  rename_all(funs(str_replace(., '_Abundance$',''))) %>% 
  filter(str_detect(pw, 'folate|pantothenate|cobalamin|biotin'))

# gather and combine with pheno info
patha <- patha %>% 
  gather(key = 'Sample', value = 'relab', names(.)[2]: names(.)[ncol(.)]) %>% 
  left_join(pheno %>% 
              select(Sample = Sampleid, CR, Toxicity), by = 'Sample')

# plot

stat.test <- compare_means(
  relab ~ pw, data = patha, group.by = "Toxicity",
  method = "wilcox.test", ref.group = 'PWY-2201: folate transformations I'
)


patha %>% 
  mutate(pw = str_replace(pw, '.+:','')) %>% 
  ggboxplot(x = 'pw', y = 'relab', color = 'Toxicity', palette  = 'lancet') +
  theme(axis.text.x = element_text(angle = 15, hjust = 1)) + 
  #stat_compare_means(aes(group = Toxicity), label = "p.signif") +
  labs(x = 'Vitamin B related pathways',
       y = 'relative abundance',
       title = 'Relative abundance in different response to Toxicity in vb related pathways') +
  ggsave('../figs/Relative abundance in different response to Toxicity in vb related pathways.jpg', width = 10, height = 7, dpi = 300)
  
```

### PCoA

```{r}
cts <- read_tsv('../data/humann2_res/humann2_final_out/humann2_genefamilies_KO_regroup_cpm_unstratified_reordered.tsv')

pheno <- read_csv('../data/cleaned_pheno/final_pheno.csv') %>% 
  dplyr::rename(sampleid = Sampleid)

plot_BC_PCoA(cts, pheno, 'Toxicity')
```

```{r}
# look at the KOs that is involved in the secondary bile acid biosynthesis

bile_KOs <- c('K01442','K15868','K15869','K15870','K15871','K15872','K15873','K15874')
bile_KOs <- paste(bile_KOs, collapse = '|')

ko_genef <- read_tsv('../data/humann2_res/humann2_genefamilies_KO_regroup.tsv') %>% 
  filter(str_detect(`# Gene Family`, bile_KOs)) %>% 
  filter(str_detect(`# Gene Family`, '\\|')) %>% 
  distinct(`# Gene Family`) %>% 
  rename(genefam = `# Gene Family`) %>% 
  separate(genefam, into = c('KO_ID','species') ,sep = '\\|'  ) 


ko_genef %>% 
  write_csv('../data/bile/bile_acid_produ_spp.csv')
```
