==> scripts/11_lefse_geneorthologs.Rmd <==
---
title: "Lefse on gene orthologs"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(tidyverse)
library(GO.db)
```

```{bash}
# ko regroup
humann2_regroup_table             --input ../data/humann2_res/humann2_genefamilies.tsv              --groups uniref90_ko              --output ../data/humann2_res/humann2_genefamilies_KO_regroup.tsv
# gropued 1+ more times:  ko 8%  eggnog 19.4% go 56.8%

# renorm to cpm
humann2_renorm_table  --input ../data/humann2_res/humann2_genefamilies_KO_regroup.tsv -u cpm -s n --output ../data/humann2_res/humann2_genefamilies_KO_regroup_cpm.tsv
#Excluding special feature: UNGROUPED)

# split table
humann2_split_stratified_table --input ../data/humann2_res/humann2_genefamilies_KO_regroup_cpm.tsv --output ../data/humann2_res/humann2_final_out
```


```{r}
pheno <- read_csv('../data/cleaned_pheno/final_pheno.csv')

# reorder the cloumn names 
source('~/pipeline/scripts/lefse/01_prepare_pheno_cts_input_table_for_lefse_run.R')

cts <- read_tsv('../data/humann2_res/humann2_final_out/humann2_genefamilies_KO_regroup_cpm_unstratified.tsv')

cts <- prepare_lefse_input_tables(pheno, cts)

cts %>% 
  write_tsv('../data/humann2_res/humann2_final_out/humann2_genefamilies_KO_regroup_cpm_unstratified_reordered.tsv')
```

```{bash}
# now use my run lefse bash script since I have already have the pheno table out 
mkdir ../data/humann2_res/lefse_ko
mkdir ../data/humann2_res/lefse_ko/tox
mkdir ../data/humann2_res/lefse_ko/CR

bash ~/projects/MSS_pipeline-/scripts/lefse/run_lefse.sh \
  ../data/lefse_input/pheno__Toxicity.txt \
  ../data/humann2_res/humann2_final_out/humann2_genefamilies_KO_regroup_cpm_unstratified_reordered.tsv \
  ../data/humann2_res/lefse_ko/tox
  
  
bash ~/projects/MSS_pipeline-/scripts/lefse/run_lefse.sh \
  ../data/lefse_input/pheno__CR.txt \
  ../data/humann2_res/humann2_final_out/humann2_genefamilies_KO_regroup_cpm_unstratified_reordered.tsv \
  ../data/humann2_res/lefse_ko/CR
```
