==> scripts/12_lefse_genefam_term_ggplot.Rmd <==
---
title: "KO & EGGNOG"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(tidyverse)
library(KEGGREST)
```

## Load the lefse results

```{r}
# load the lefse res of tox
# the input : lefse res files
fns <- c(list.files('../data/humann2_res/lefse_ko/', pattern = '*/*.res$', full.names = T),
         list.files('../data/humann2_res/lefse_ko/', pattern = '*/*.res$', full.names = T))


# join all of the tables feature together
feature <- fns %>% 
  set_names(fns_name) %>% 
  # when meeting parsing failures, you need to specify all column types.
  map(~ read_tsv(., col_names = c('genefam','xx','direction','score','pval'),  col_types = 'ccccc') %>% 
        filter(!is.na(score))) %>% 
  bind_rows(.id = 'group') %>% 
  mutate(group = str_replace(group, 'genefam_abundace_pheno__','')) %>% 
  mutate(group = str_replace(group, '.res$','')) %>% 
  select(-xx, -pval) %>% 
  mutate(KOID = str_replace(genefam, '\\.g.+$','')) 
 


# feature %>% 
#   write_csv('../data/lefse_genefam/with_unmapped_sig_feature.csv')


```

## Use KEGGREST pkg to get the actual term

### work on the Toxicity first

```{r}
# since this pkg only retrieve 10 records one time, so need to split the df.
tox_df <- feature %>% 
  filter(group == 'Toxicity') 

tox_ko <- tox_df %>% 
  split(cut_width(1:nrow(.), 10, boundary=0)) %>% 
  map(~ pull(.data = ., KOID) %>% 
        map(~ list(KOID = keggGet(.)[[1]]$ENTRY,
                   KO_name = keggGet(.)[[1]]$NAME,
                   KO_def = keggGet(.)[[1]]$DEFINITION)) %>% 
        bind_rows()) %>% 
  bind_rows() %>% 
  distinct(KOID, .keep_all = T)
 
# in the table some have duplicated KOID since several terms have a trailing species name
tox_lefse <- tox_df %>% 
  left_join(tox_ko, by = 'KOID')

```

```{r}
# format to visualize in lefse style
# change the "N" direction to be minus score
tox_final <- bind_rows(
  tox_lefse %>%
    split(.$direction) %>%
    pluck('Toxicity_N') %>%
    mutate(score = -as.numeric(score)),
  tox_lefse %>%
    split(.$direction) %>%
    pluck('Toxicity_Y') %>%
    mutate(score = as.numeric(score))
) %>% 
  # wanna highlight there are some terms have trailing species
  mutate(have_species = if_else(str_detect(genefam, '\\.g__.+$'), str_extract(genefam, 'g__.+$'), ''))   %>% 
  mutate(have_FP = if_else(have_species == 'g__Faecalibacterium_s__Faecalibacterium_prausnitzii', 'YES', 'NO')) %>% 
  # combine the name and def to be a new var
  mutate(KO_term = str_glue('{KO_name}: {KO_def}'))
  

# for plotting purpose plotting the unique KO_term rows 
ggplot(data = tox_final %>% 
         distinct(KO_term, .keep_all = T) , aes(x = reorder(KO_term, score), y = score, fill = direction, color = have_FP)) +
  geom_bar(stat='identity') +
  coord_flip() +
  scale_color_manual(values = c('blue', 'yellow'))+
  scale_fill_manual(values = c('#EC0000', '#42B440')) +
  theme_minimal() +
  theme(axis.title.y  = element_blank(), 
        legend.position="bottom") +
  labs(title = str_glue('Response to class : Toxicity') ,
       y = 'Score') + 
  ggsave(str_glue('/Volumes/castoricenter/CAR-Tcell_and_Microbiota/humann2_shotgun/figs/KEGG_ortholog_lefse_plot/Response to class: Toxicity.pdf'), device = 'pdf', width = 18, height = 10, dpi = 300)
 
```
