==> scripts/13_PCoA_brayCurtis_KO_abundance.Rmd <==
---
title: "PCoA with bray curtis on KO abundance"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)  
library(tidyverse)
library(ggpubr)
library(vegan)
```

Attempt on making the figure 7 A and B of paper https://www.nature.com/articles/srep34826
Making PCoA plot with Bray-Curtis distance on the KO abundace matrix

```{r}
# using the unstratified cts first
pheno <- read_csv('../data/current_pheno.csv') %>% 
  filter(! Sampleid %in% c("1725E", "2222A" ,"2245A"))

ko <- read_tsv('../data/humann2_res/genefamily_split_KO/humann2_genefamilies_KO_regroup_relab_unstratified.tsv') %>% 
  rename(KO = names(.)[1]) %>% 
  rename_all(funs(str_replace(., '_Abundance-RPKs$',''))) 
  

ko <- bind_cols(ko[,1],ko[,pheno$Sampleid]) 
all.equal(pheno$Sampleid, colnames(ko)[2:ncol(ko)])

# turn the ko table in to a matrix
ko_matrix <- ko %>% 
  column_to_rownames('KO') %>% 
  as.matrix()
```

## Looking at the whole table

```{r}
pw <- read_tsv('../data/humann2_res/humann2_final_out/humann2_pathabundance_relab_unstratified.tsv') %>% 
  rename(pw = names(.)[1]) %>% 
  rename_all(funs(str_replace(., '_Abundance$',''))) 

pw <- bind_cols(pw[,1],pw[,pheno$Sampleid]) 
all.equal(pheno$Sampleid, colnames(pw)[2:ncol(pw)])

pw_matrix <- pw %>% 
  column_to_rownames('pw') %>% 
  as.matrix()
```

```{r}
# PCoA
pheno_col <- pheno %>% 
  mutate(tox_col = if_else(Toxicity == 'Y', 'red', 'green '),
         CR_col = if_else(CR == 'Y', 'red', 'green '))

bc_dist <-  vegdist(t(pw_matrix), "bray")
pc_bc <- cmdscale(bc_dist, k = 2)
plot(pc_bc[,1], pc_bc[,2], col = pheno_col$tox_col, main = 'Pathway abundance color by tox')
plot(pc_bc[,1], pc_bc[,2], col = pheno_col$CR_col, main = 'Pathway abundance color by CR')
```

```{r}
# do the PcoA with bray curtis distance  
bc_dist <-  vegdist(t(ko_matrix), "bray")
pc_bc <- cmdscale(bc_dist, k = 2)
plot(pc_bc[,1], pc_bc[,2], col = pheno_col$tox_col, main = 'KO abundance color by tox')
plot(pc_bc[,1], pc_bc[,2], col = pheno_col$CR_col, main = 'KO abundance color by CR')
```

It just doesn't separate whether using gene ortholog or pathway abundance.       

## Looking at subset of pathways

```{r}
vb_matrix <- pw %>% 
  filter(str_detect(pw, 'folate|pantothenate|cobalamin|biotin')) %>% 
  column_to_rownames('pw') %>% 
  as.matrix()


bc_dist <-  vegdist(t(vb_matrix), "bray")
pc_bc <- cmdscale(bc_dist, k = 2)
plot(pc_bc[,1], pc_bc[,2], col = pheno_col$tox_col, main = 'Pathway abundance color by tox')
plot(pc_bc[,1], pc_bc[,2], col = pheno_col$CR_col, main = 'Pathway abundance color by CR')

```


#############################################################################

```{r}
# the PCoA of the species abundance table
spp <- read_tsv('../data/metaphlan2_result/merged_abundance_table_species.txt') %>% 
  rename_all(funs(str_replace(., '_profile$','')))

pheno <- read_csv('../data/cleaned_pheno/final_pheno.csv') %>% 
  select(-CAR_Group, -Censor_reason ) %>% 
  mutate_at(vars(str_which(names(.),'Sampleid', negate = T)), factor)

# to rearrange the cts table 
spp <- bind_cols(spp[,1],spp[,pheno$Sampleid]) 

# to put the feature table in the sample as rowname format to do the ML 
spp <- spp %>% 
  column_to_rownames('ID') %>% 
  as.matrix()
 
all.equal(colnames(spp), pheno$Sampleid)
```

```{r}
pheno_col <- pheno %>% 
  mutate(tox_col = if_else(Toxicity == 'Y', 'red', 'green '),
         CR_col = if_else(CR == 'Y', 'red', 'green '))

bc_dist <-  vegdist(t(spp), "bray")
pc_bc <- cmdscale(bc_dist, k = 2) 
plot(pc_bc[,1], pc_bc[,2], col = pheno_col$tox_col, main = 'species abundance color by tox')
plot(pc_bc[,1], pc_bc[,2], col = pheno_col$CR_col, main = 'species abundance color by CR')


# updated pheno
uppheno <- read_csv('../data/cleaned_pheno/pheno_w_expanded_tox.csv', col_types = 'cfffffff')

# ggplot 
pc_bc %>% 
  as.data.frame() %>% 
  rownames_to_column('Sampleid') %>% 
  full_join(uppheno, by = 'Sampleid') %>% 
  ggplot(aes(x  = V1, y = V2, color =  Toxicity, fill = Toxicity)) +
  geom_point() +
  labs(x = 'PCo 1', y = 'PCo 2', title = 'PCoA plot for response to Toxicity') +
  theme_classic() +
  scale_color_manual(values = c('#42B440', '#EC0000'))  +
  scale_fill_manual(values = c('#42B440', '#EC0000')) +
  ggsave('../figs/PCoA/response_tox.jpg', dpi = 300, width = 6, height = 5)

 


pc_bc %>% 
  as.data.frame() %>% 
  rownames_to_column('Sampleid') %>% 
  full_join(uppheno, by = 'Sampleid') %>% 
  ggplot(aes(x  = V1, y = V2, color =  CR, fill = CR)) +
  geom_point() +
  labs(x = 'PCo 1', y = 'PCo 2', title = 'PCoA plot for response to CR') +
  theme_classic() +
  scale_color_manual(values = c('#42B440', '#EC0000'))  +
  scale_fill_manual(values = c('#42B440', '#EC0000')) +
  ggsave('../figs/PCoA/response_CR.jpg', dpi = 300, width = 6, height = 5) 




pc_bc %>% 
  as.data.frame() %>% 
  rownames_to_column('Sampleid') %>% 
  full_join(uppheno, by = 'Sampleid') %>% 
  ggplot(aes(x  = V1, y = V2, color =  ORR, fill = ORR)) +
  geom_point() +
  labs(x = 'PCo 1', y = 'PCo 2', title = 'PCoA plot for response to ORR') +
  theme_classic() +
  scale_color_manual(values = c('#42B440', '#EC0000'))  +
  scale_fill_manual(values = c('#42B440', '#EC0000')) +
  ggsave('../figs/PCoA/response_ORR.jpg', dpi = 300, width = 6, height = 5) 
```
