==> scripts/15_RF_find_top_features.Rmd <==
---
title: "Random Forest to find the top important features"
author: "Anqi Dai"
date: "`r Sys.Date()`"
output:
  html_document:
    toc: true
    toc_float: true
    toc_depth: 3 
    code_folding: hide
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, message = F, warning = F)
library(randomForest)
library(ggpubr)
library(tidyverse)
library(kableExtra)
library(arrangements)
```

## Response to Toxicity

### Glimpse at pheno table

```{r}
# the species table from metaphlan2
# the species abundance here are actually in percent
s_tb <- read_tsv('../data/metaphlan2_result/merged_abundance_table_species.txt') %>% 
  rename_all(funs(str_replace(., '_profile$','')))

s_tb_sum <- s_tb %>% 
  column_to_rownames('ID') %>% 
  as.matrix() %>% 
  apply(., 2, sum)

# load the pheno
pheno <- read_csv('../data/cleaned_pheno/final_pheno.csv') %>% 
  select(-CAR_Group, -Censor_reason ) %>% 
  mutate_at(vars(str_which(names(.),'Sampleid', negate = T)), factor)

# to rearrange the cts table 
s_tb <- bind_cols(s_tb[,1],s_tb[,pheno$Sampleid]) 

# write out the reordered version
s_tb %>% 
  write_tsv('../data/metaphlan2_result/merged_abundance_table_species_reordered.tsv')

# to put the feature table in the sample as rowname format to do the ML 
s_tb_rf <- s_tb %>% 
  column_to_rownames('ID') %>% 
  as.matrix() %>% 
  t

# check the sample name really match
all.equal(rownames(s_tb_rf), pheno$Sampleid)

# pheno %>% 
#   kable(caption = 'Phenotype info of the final 33 baseline sample patients') %>% 
#   kable_styling(full_width = F)

pheno  %>% 
  summary 
```

```{r}
# check whether the pheno vars are correlated (in pairs)
all_var <- names(pheno)
combi <- combinations(all_var[all_var != 'Sampleid'], 2) %>% 
  as.data.frame(stringsAsFactors = F)

phe_cor <- apply(combi, 1, function(Row) {
  phe_fil = pheno %>% 
    select(Row[['V1']], Row[['V2']]) %>% 
    table()
  
  chisq.test(phe_fil)$p.value
}) %>% 
  as.data.frame() %>% 
  bind_cols(combi) %>% 
  mutate(phe_var = str_glue('{V1}__{V2}')) %>% 
  select(-V1, -V2) %>% 
  rename(corrected_pval = names(.)[1]) %>% 
  arrange(corrected_pval) %>% 
  mutate(CORR = if_else(corrected_pval < 0.1, 'Yes', 'No'))


phe_cor %>% 
  kable(caption = 'Test on independence between two of the pheno vars') %>% 
  kable_styling(full_width = F, position = 'left')
```



### Use all features as predictors

```{r}
tox_RF <- randomForest(s_tb_rf, factor(pheno$Toxicity), importance = T)

impor_feature <- importance(tox_RF, type=1) %>% 
  as.data.frame() %>% 
  rownames_to_column(var = 'species') %>% 
  arrange(MeanDecreaseAccuracy)


# plot the err rate change
tox_RF  %>% 
  pluck('err.rate') %>% 
  as.data.frame() %>% 
  mutate(ntree = seq(1, nrow(.))) %>% 
  gather(key = 'type', value = 'ERR', OOB:Y) %>% 
  ggline(x = 'ntree', y = 'ERR', color = 'type', palette = 'lancet') +
  labs(x = 'Number of trees', 
       y = 'Error rate')

tox_RF
```

Basically assigning everything to Y. 

### Use all features + uncorrelated pheno var as predictors

Uncorrelated pheno var: ORR, Disease_Status_CAR for Toxicity

```{r} 
# bind the pheno with the transposed cts matrix
combined <- cbind(s_tb_rf, pheno %>% 
                    select(Toxicity, ORR, Disease_Status_CAR)) 

combined_tox_RF <- randomForest(formula = Toxicity ~ ., data = combined,  importance = T )  

combined_tox_RF  %>% 
  pluck('err.rate') %>% 
  as.data.frame() %>% 
  mutate(ntree = seq(1, nrow(.))) %>% 
  gather(key = 'type', value = 'ERR', OOB:Y) %>% 
  ggline(x = 'ntree', y = 'ERR', color = 'type', palette = 'lancet') +
  labs(x = 'Number of trees', 
       y = 'Error rate')

combined_tox_RF
```


### Use more abundant features(species abundance > 1%) + uncorrelated pheno var as predictors

```{r}
# check the dist of the species relab sum of the table
s_sum <- s_tb_rf %>% 
  apply(., 2, sum)

data_frame(species = names(s_sum),
           relab_sum = s_sum) %>% 
  gghistogram(x = 'relab_sum', bins = 50)

# remove the species that have less than 1(%) abundance 
length(s_sum[s_sum > 1]) 


combined_fil <- cbind(s_tb_rf[,names(s_sum[s_sum > 1])], pheno %>% 
                    select(Toxicity, ORR, Disease_Status_CAR)) 


combined_tox_RF_fil <- randomForest(formula = Toxicity ~ ., data = combined_fil,  importance = T , ntree = 1000 )  

combined_tox_RF_fil  %>% 
  pluck('err.rate') %>% 
  as.data.frame() %>% 
  mutate(ntree = seq(1, nrow(.))) %>% 
  gather(key = 'type', value = 'ERR', OOB:Y) %>% 
  ggline(x = 'ntree', y = 'ERR', color = 'type', palette = 'lancet') +
  labs(x = 'Number of trees', 
       y = 'Error rate')

combined_tox_RF_fil
```

No improvements at all. I think because the relatively small sample size and the signal itself is weak. There's no guarantee that there is association between microbiome and the outcome to begin with. 

### Use top N fraction features + uncorrelated pheno var as predictors


```{r}
# to use the top N percent features

# to test on different threshold of selecting most abundant species
sum_thre <- seq(0.5, 1, 0.05)
tree_num <- 1000

res <- sum_thre %>% 
  set_names(sum_thre) %>% 
  map(function(thre) {
    s_sum_top <- quantile(s_sum, thre)
    combined_fil <- cbind(s_tb_rf[,names(s_sum[s_sum > s_sum_top])], pheno %>% 
                    select(Toxicity, ORR, Disease_Status_CAR)) 
    
    tibble(
      num_fea = length(s_sum[s_sum > s_sum_top]) + 2,
      OOB_err_rate = randomForest(formula = Toxicity ~ ., data = combined_fil,  importance = T , ntree = tree_num )[['err.rate']][tree_num, 1]
      
    ) 
  }) %>% 
  bind_rows(.id = 'thre')


res %>% 
  kable(caption = 'Top N fraction features and the OOB error rate') %>% 
  kable_styling(full_width = F, position = 'left')
```


### See the important features(species)

Using the table with features abundance > 1% and the two uncorrelated vars. 

```{r}
# repeat the RF M = 100 times
# pull top 10 features
N <- 10 
M <- 100 
pull_top_features <- function(){
  importance(randomForest(formula = Toxicity ~ ., data = combined_fil,  importance = T ) , type=1, scale = F) %>% 
    as.data.frame() %>% 
    rownames_to_column(var = 'species') %>% 
    arrange(desc(MeanDecreaseAccuracy)) %>% 
    top_n(N) %>% 
    pull(species)
}

res <- replicate(M, pull_top_features())

stable_top_fea <- table(res) %>% 
  as.data.frame() %>% 
  rename(species = names(.)[1]) %>% 
  arrange(desc(Freq))

stable_top_fea %>% 
  kable(caption = 'Frequency of features shown as top 10 in 100 random forest runs') %>% 
  kable_styling(full_width = F, position = 'left')

```


What's interesting is this species *Ruminococcus gnavus*. [This recent paper](https://www.pnas.org/content/116/26/12672) points out that *Ruminococcus gnavus* produces a polysaccharide molecule that induces the production of inflammatory cytokines.

And [this paper](https://www.ncbi.nlm.nih.gov/pubmed/29330424) mentions that "*Faecalibacterium prausnitzii* decrease and *Ruminococcus gnavus* increase were the essential characteristics in CHF patients' gut microbiota".

### Use the toxicity grade

```{r}
crs <- read_csv('../data/cleaned_pheno/cleaned_CRS.csv')
 
pheno <- pheno %>% 
  full_join(crs %>% 
              select(Sampleid  = sampleid, CRS_grade), by = 'Sampleid') %>% 
  mutate(CRS_grade = factor(CRS_grade))

all.equal(rownames(s_tb_rf), pheno$Sampleid)

########################################################################################

# what if just do as Tsoni says that classify 0 and 1 as N the other as Y
pheno <- pheno %>% 
  mutate(tox_crs = factor(if_else(as.character(CRS_grade) %in% c('0','1'), 'NO', 'YES')))


# write out this pheno for future use

pheno %>% 
  write_csv('../data/cleaned_pheno/pheno_w_expanded_tox.csv')

# select the 18 features that have > 95% relab
s_sum_top <- quantile(s_sum, 0.95)
combined_fil <- cbind(s_tb_rf[,names(s_sum[s_sum > s_sum_top])], pheno %>% 
                    select(tox_crs, ORR, Disease_Status_CAR)) 


dim(combined_fil) 
   


CRS_RF <- randomForest(formula = tox_crs ~ ., data = combined_fil,  importance = T )  

CRS_RF  %>% 
  pluck('err.rate') %>% 
  as.data.frame() %>% 
  mutate(ntree = seq(1, nrow(.))) %>% 
  gather(key = 'type', value = 'ERR', 'OOB':'YES') %>% 
  ggline(x = 'ntree', y = 'ERR', color = 'type', palette = 'lancet') +
  labs(x = 'Number of trees', 
       y = 'Error rate') 

CRS_RF    # very high error rate. not working. 
```
