==> scripts/16_RF_pathway.Rmd <==
---
title: "RF pathway abundance"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(randomForest)
library(ggpubr)
library(tidyverse)
library(kableExtra)
```

```{r}
pheno <- read_csv('../data/cleaned_pheno/final_pheno.csv') 

pw <- read_tsv('../data/humann2_res/humann2_final_out/humann2_pathabundance_cpm_unstratified.tsv') %>% 
  rename(pathway = names(.)[1]) %>% 
  rename_all(funs(str_replace(., '_Abundance$',''))) %>% 
  # remove the unmapped and the unintegrated record
  filter(!pathway %in% c('UNMAPPED','UNINTEGRATED'))

pw <- bind_cols(pw[,1],pw[,pheno$Sampleid]) 


pw_rf <- pw %>% 
  column_to_rownames('pathway') %>% 
  as.matrix() %>% 
  t

all.equal(rownames(pw_rf), pheno$Sampleid)


tox_RF_PW <- randomForest(pw_rf, factor(pheno$Toxicity), importance = T)
```
