==> scripts/18_lefse_metaphlan2_CR_tox.Rmd <==
---
title: "Format for lefse and run on the classes CR and tox"
output: html_document
---
         
Melody's project

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(tidyverse)
```

# Run unstratified pathway abundance of Melody's project thru lefse
* CR & Toxicity

## Sorting out the pheno table for the valid samples

```{r}
pheno <- read_csv('../data/cleaned_pheno/final_pheno.csv') %>% 
  select(-CAR_Group, -Censor_reason ) 
```

## Creating the class+subclass combi
 
```{r}
# get different combination of the classes
# using no subclass (11-4-2019)

classes <- c('CR', 'Toxicity')


combi <- expand.grid(classes = classes,subclass = subclass)

# use each combi row in the above df to create the pheno table for the lefse input

apply(combi, 1, function(row) {
  pheno %>% 
    select(row[['classes']]) %>% 
    t() %>% 
    write.table(str_glue('../data/lefse_input/pheno__{row[["classes"]]}.txt'), sep = '\t', quote = F, row.names = T, col.names = F)
})
```
   
## Sorting the relab abundance table so that sampleid would align  
  
```{r}
# keeping the leading taxa info 
spp <- read_tsv('../data/metaphlan2_result/merged_abundance_table.txt') %>% 
  rename_all(funs(str_replace(., '_profile$','')))

spp <- bind_cols(spp[,1],spp[,pheno$Sampleid]) 

all.equal(colnames(spp)[2:ncol(spp)], pheno$Sampleid)

# write out this table 
spp %>% 
  write_tsv('../data/metaphlan2_result/merged_abundance_table_WHOLE_reordered.tsv')
```

## Create the bash commmand to run the lefse steps

```{r}
# create the commad in df with str_glue and then output to a bash script
pheno_combi <- tibble(pheno_file = list.files('../data/lefse_input', pattern = '[Toxicity|CR].txt$', full.names = T)) %>% 
  mutate(combi = str_extract(pheno_file, 'pheno.+\\.')) 
 
pheno_combi %>% 
  mutate(cat_cmd = str_glue('cat {pheno_file} ../data/metaphlan2_result/merged_abundance_table_WHOLE_reordered.tsv > ../data/lefse_input/spp_abundace_{combi}tsv')) %>% 
  select(cat_cmd) %>% 
  write_csv('cat_combi_spp.sh', col_names = F)
```


```{r}
# create the cmd to do the format input stuff
pheno_combi %>% 
  mutate(input_tsv = str_glue('../data/lefse_input/spp_abundace_{combi}tsv'),
         fmt_input = str_replace(input_tsv, 'tsv','in'),
         fmt_cmd = str_glue('format_input.py {input_tsv} {fmt_input} -c 1 -u 2 -o 1000000')) %>% 
  select(fmt_cmd) %>% 
  write_csv('format_input_spp.sh', col_names = F)
```


```{r}
# the actual step to run lefse
pheno_combi %>% 
  mutate(fmt_input = str_glue('../data/lefse_input/spp_abundace_{combi}in'),
         lefse_res = str_replace(fmt_input, 'in$','res'),
         run_cmd = str_glue('run_lefse.py {fmt_input} {lefse_res}')) %>% 
  select(run_cmd) %>% 
  write_csv('run_lefse_spp.sh', col_names = F)
```


```{r}
# plotting cmd 
pheno_combi %>% 
  mutate(lefse_res = str_glue('../data/lefse_input/spp_abundace_{combi}res'),
         fea_png = str_glue('../figs/spp_abundace_{combi}png'),
         cmd = str_glue('plot_res.py {lefse_res}  {fea_png} --feature_font_size 4 --width 10 --dpi 300 --title {combi}')) %>% 
  select(cmd) %>% 
  write_csv('plotting_de_feature_spp.sh', col_names = F)
```


```{r}
# plot the cladogram   
pheno_combi %>% 
  mutate(lefse_res = str_glue('../data/lefse_input/spp_abundace_{combi}res'),
         cla_png = str_glue('../figs/cladogram/spp_abundace_lefse_{combi}png'),
         cmd = str_glue('plot_cladogram.py {lefse_res}  {cla_png}  --dpi 300 --format png --title {combi} --label_font_size 4')) %>% 
  select(cmd) %>% 
  write_csv('plotting_cladogram_spp.sh', col_names = F)
```



####################################### below is command line ##########################################

## Run in terminal the below bash script

```{bash}
# this is the bash script to do the actual concatenation 
bash cat_combi_spp.sh
bash format_input_spp.sh
bash run_lefse_spp.sh
bash plotting_de_feature_spp.sh
bash plotting_cladogram_spp.sh
```
