==> scripts/21_correlation_genus_bw_16S_shotgun.Rmd <==
---
title: "Correlation between 16S and shotgun"
author: "Anqi Dai"
date: "`r Sys.Date()`"
output:
  html_document:
    toc: true
    toc_float: true
    toc_depth: 3
    code_folding: hide
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, warning = F, message = F)
library(tidyverse)
library(ggpubr)
library(RColorBrewer)
library(kableExtra)
```

# At the genus level

## Load the two datasets

```{r}
# shotgun
gun <- read_tsv('../data/metaphlan2_result/merged_abundance_table_genus.txt') %>% 
  rename_all(funs(str_replace(., '_profile$',''))) %>% 
  rename(genus = ID)

pheno <- read_csv('../data/cleaned_pheno/final_pheno.csv') %>% 
  select(-CAR_Group, -Censor_reason )
```


```{r}
dada2 <- read_rds('/Volumes/castoricenter/CAR-Tcell_and_Microbiota/data/asv_set.RDS')


# how many samples are in the dada2 data
dada2_nsamps <- table(dada2 %>% 
  pluck('asv_counts') %>% 
    pull(sampleid)) %>% 
  names %>% 
  length

# get the relab at phylum level
sixteen <- dada2 %>% 
  pluck('asv_counts') %>% 
  select(ASV, count_relative, sampleid) %>% 
  left_join(dada2 %>% 
    pluck('asv_annotation') %>% 
    select(ASV, genus ), by = 'ASV') %>% 
  filter(!is.na(genus)) %>% 
  # select the 33 that are in the shotgun
  filter(sampleid %in% pheno$Sampleid) %>% 
  group_by(sampleid, genus) %>% 
  summarise(relab = sum(count_relative)) %>% 
  spread(key = 'sampleid', value = 'relab', fill = 0) 


```


```{r}
# same genus term
ol_genus <- intersect(sixteen$genus, gun$genus)

# have an idea that i better put the two of them table in the same table with overlapped phylum terms

two <- list(shotgun = gun,
            s16S = sixteen) %>% 
  imap(~  ..1 %>% 
      filter(genus %in% ol_genus) %>% 
      gather(key = 'sampleid', value = 'relab', names(.)[2] : names(.)[ncol(.)]) %>% 
      mutate(Group = ..2) %>% 
      mutate_at(vars(str_which(names(.),'relab', negate = T)), factor)
  ) 

Two <- two %>% 
  pluck('shotgun') %>% 
  rename(shotgun_relab = relab) %>% 
  select(-Group) %>% 
    full_join(two %>% 
    pluck('s16S') %>% 
    rename(s16S_relab = relab) %>% 
    select(-Group)) %>%  
  # to uniform the two data
  mutate(s16S_relab = s16S_relab*100) %>% 
  # log10 transform the data add pseudocounto f 0.000001
  mutate(shotgun_relab = log10(shotgun_relab + 0.0001),
         s16S_relab = log10(s16S_relab + 0.0001))
  
```

## Compare with Spearman correlation for each sample

```{r}
Two %>% 
  ggplot(aes(x = shotgun_relab, y = s16S_relab)) +
    geom_point() +
    facet_wrap(~sampleid) +
  geom_abline(intercept = 0, slope = 1, col = 'pink', linetype = 'dashed') +
  geom_vline(xintercept = 0, col= 'forestgreen', linetype = 'dashed') +
  geom_hline(yintercept = 0, col= 'forestgreen', linetype = 'dashed') +
  labs(x = 'log10(Shotgun data)',
       y = 'log10(16S data)',
       title = 'log10 transformed relative abundance comparison between shotgun and 16S data\nat genus level') +
  theme_light()+
  ggsave('../figs/correlation/cor_between_16_and_shotgun_genus.jpg', width = 10, height = 10, dpi = 300)


# plot a single sample's plot to show the details
Two %>% 
  filter(sampleid == '275G') %>% 
  ggplot(aes(x = shotgun_relab, y = s16S_relab)) +
    geom_point() +
    geom_abline(intercept = 0, slope = 1, col = 'pink', linetype = 'dashed') +
    geom_vline(xintercept = 0, col= 'forestgreen', linetype = 'dashed') +
    geom_hline(yintercept = 0, col= 'forestgreen', linetype = 'dashed') +
    labs(x = 'log10(Shotgun data)',
         y = 'log10(16S data)',
         title = 'log10(relative abundance) comparison (genus): 275G') +
    theme_light()+
    ggsave('../figs/correlation/275G_cor_between_16_and_shotgun_genus.jpg', width = 5, height = 5, dpi = 300)


# spearman_cor
spearman_cor <- Two %>% 
  split(.$sampleid) %>% 
  map_dfr(~ round(cor.test(.$shotgun_relab, 
                           .$s16S_relab, 
                           method = 'spearman', 
                           exact = F)$estimate, 2)) %>% 
  gather(key = 'sampleid', value = 'rho') %>% 
  arrange(rho)


spearman_cor %>% 
  kable(caption = 'Spearman correlation at genus level') %>% 
  kable_styling(full_width = F, position = 'left')


pearson_cor <- Two %>% 
  split(.$sampleid) %>% 
  map_dfr(~ round(cor.test(.$shotgun_relab, 
                           .$s16S_relab, 
                           method = 'pearson', 
                           exact = F)$estimate, 2)) %>% 
  gather(key = 'sampleid', value = 'cor') %>% 
  arrange(cor)  
  
```
 

```{r}
py_cor <- read_csv('../data/spearman_cor_phylum.csv') %>% 
  rename(pylum = rho)

py_cor %>% 
  full_join(spearman_cor %>% 
              rename(genus = rho), by = 'sampleid') %>% 
  gather(key = 'Group', value = 'Cor', names(.)[2]: names(.)[ncol(.)]) %>% 
  ggboxplot(x = 'Group', y = 'Cor',  title = 'Comparison of spearman correlation') +
  stat_compare_means() +
  ggsave('../figs/pylum_genus_correlation.jpg')



```

```{r}
py_cor <- read_csv('../data/pearson_cor_phylum.csv') %>% 
  rename(pylum = cor)

py_cor %>% 
  full_join(pearson_cor %>% 
              rename(genus = cor), by = 'sampleid') %>% 
  gather(key = 'Group', value = 'Cor', names(.)[2]: names(.)[ncol(.)]) %>% 
  ggboxplot(x = 'Group', y = 'Cor',  title = 'Comparison of Pearson correlation') +
  stat_compare_means() +
  ggsave('../figs/pylum_genus_correlation_pearson.jpg')
```
