==> scripts/27_maASlin2.Rmd <==
---
title: "maASlin2"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(Maaslin2) 
library(eeptools)
library(tidyverse)
library(arrangements)
```

```{r}
# get the MRN of my target samples 
link <- read_excel('/Volumes/castoricenter/CAR-Tcell_and_Microbiota/CTC Event Table 12122019.xlsx',
                    sheet = 4) %>% 
  dplyr::select(pt_id, 
                MRN,
                disease_sta = `Key: Patient Disease Status at CAR`)


# format the whole table column and keep the samples in the final cohort 33
have_sg <- read_csv('../data/baselinesforAngel10032019.csv') %>% 
  filter(shotgun_YN == 'yes') %>% 
  # removig 1943A cuz its low quality
  filter(sample_id != '1943A') %>% 
  ### FUCK THIS TABLE OF ../data/baselinesforAngel10032019.csv IS WRONG IN 161, it didn’t pass QC because not enough DNA material was present. SO WE DON'T HAVE THE SHOTGUN DATA.
  filter(!sample_id %in% c('161'))

# now I have MRN in the below tble
phe_M <- read_excel('/Volumes/castoricenter/CAR-Tcell_and_Microbiota/CTC Event Table 11252019 wo Censor .xlsx',
                    sheet = 5) %>% 
  rename_all(funs(str_replace_all(., ' \\(.+$','') %>% 
  str_replace_all(., ' ','_') %>% 
  str_replace(., ':|-','_'))) %>% 
  filter(is.na(Censor_reason)) %>% 
  filter(Baseline_Sample_ID %in% have_sg$sample_id) %>% 
  left_join(link, by = 'pt_id')


# sub selecting the vars I'm interested in 
pheno <- phe_M %>% 
  dplyr::select(MRN,
                sampleid = Baseline_Sample_ID,
                CR, 
                ORR, 
                Toxicity, 
                Cytokine_Release_Syndrome,
                ICANS,
                Antibiotics_4_weeks_post_CAR_T_cells,
                Disease_Class,
                disease_sta
                )
```

```{r}
# add the lines of therapy var
lines <- read_excel('/Volumes/castoricenter/CAR-Tcell_and_Microbiota/CTC Event Table 11252019 wo Censor .xlsx',
                    sheet = 8)  %>% 
  rename_all(funs(str_replace_all(., ' ','_'))) %>% 
  dplyr::select(MRN, Prior_lines_of_therapy)
```


```{r}
# get the clinical data info  from ying pulled data
load('/Volumes/castoricenter/CAR-Tcell_and_Microbiota/YingRCode/ctc_melody.RData')

# load the table I made for the abx exposure
abx <- read_csv('../data/cleaned_pheno/shotgun_cohort_abx_4w_baseline_expo_table.csv')
```

```{r}
# get the specifically the three ILs, ferriton and crp

IL_table <- ctc.interleukin %>% 
  # remove that one record that has serum in subtest_name
  filter(!str_detect(subtest_name, 'Serum')) %>% 
  # format the 1,000 such value to remove the comma
  mutate(result_text = str_replace(result_text, ',',''),
         result_text = if_else(str_detect(result_text, '^<'), '0', result_text)) %>% 
  mutate(result_text = as.numeric(result_text)) %>% 
  group_by(MRN, subtest_name) %>% 
  summarise(IL =  max(result_text)) %>% 
  mutate(subtest_name = str_replace(subtest_name, 'Interleukin','IL'),
         subtest_name= str_replace_all(subtest_name, ' ','_')) %>% 
  spread(key = 'subtest_name' , value = 'IL')

# the three ILs, select the highest value if there are multiple 
pheno <- pheno %>% 
  # get patient info such as age and sex (using MRN)
  left_join(ctc.patients %>% 
              dplyr::select(MRN, sex, dob), by = 'MRN') %>% 
  # get the hightest ferritin value 
  left_join(ctc.ferritin %>% 
              dplyr::select(MRN, result_text) %>% 
              group_by(MRN) %>% 
              summarise(ferritin = max(result_text)), by = 'MRN') %>% 
  left_join(ctc.crp %>% 
              dplyr::select(MRN, result_text) %>% 
              mutate(result_text = if_else(str_detect(result_text, '^<'), '0', result_text),
                     result_text = as.numeric(result_text)) %>% 
              group_by(MRN) %>% 
              summarise(crp  = max(result_text, na.rm = T)), by = 'MRN') %>% 
  left_join(IL_table, by  ='MRN') %>% 
  left_join(lines, by = 'MRN') %>% 
  left_join(abx, by = 'MRN')
  

pheno <- pheno %>% 
  mutate(age = round(age_calc(pheno$dob, enddate = Sys.Date(), units = "years", precise = TRUE),0)) %>% 
  dplyr::select(-dob)

  
pheno %>% 
  write_csv('../data/cleaned_pheno/final_pheno.csv')
```


```{r pheno_final}
# now change the columns to different type: dbl or factor
pheno_final  <- pheno %>%  
  # change the ferritin to dbl
  mutate(ferritin = as.numeric(ferritin),
         disease_sta = as.character(disease_sta)) %>% 
  # change all the pheno vars to factors
  mutate_if(is.character, as.factor) %>% 
  mutate(MRN = as.character(MRN),
         sampleid = as.character(sampleid)) %>% 
  #mutate(IL_6 = log(IL_6),
  #       IL_10 = log(IL_10)) %>% 
  column_to_rownames('sampleid')  


colnames(pheno_final)


pheno_final %>% 
  summary
```

```{r}
# check the correlation between the pheno vars that are factors
all_var <- pheno_final %>% 
  select_if(is.factor) %>% 
  names
  
  
combi <- combinations(all_var, 2) %>% 
  as.data.frame(stringsAsFactors = F)

phe_cor <- apply(combi, 1, function(Row) {
  phe_fil = pheno_final %>% 
    dplyr::select(Row[['V1']], Row[['V2']]) 
  
  round(cor.test(pheno_final[,Row[['V1']]],  pheno_final[,Row[['V2']]],method = 'spearman', 
                           exact = F)$estimate, 2)
  #chisq.test(phe_fil)$p.value
}) %>% 
  as.data.frame() %>% 
  bind_cols(combi) %>% 
  mutate(phe_var = str_glue('{V1}__{V2}')) %>% 
  dplyr::select(-V1, -V2) 
```

```{r}
pw <- read_tsv('../data/humann2_res/humann2_final_out/humann2_pathabundance_cpm_unstratified.tsv') %>% 
  dplyr::rename_all(funs(str_replace(., '_Abundance$',''))) %>% 
  dplyr::rename(pw = names(.)[1])  %>% 
  # remove the things after the : so that its easier to see on the plot(the pathway name)
  mutate(pw = str_replace(pw, ':.+$','')) %>% 
  column_to_rownames('pw') %>% 
  as.matrix() %>% 
  t() %>% 
  as.data.frame()

# filter on the above matrix to include pathways existing in at least 20% of the samples (>=4)
samp_num <- nrow(pw)

ok_samps <- colSums(pw != 0)

# to see how many samps pass the number of zeros threshold
pers <- seq(0.1, 0.99, 0.1)

pers %>% 
  set_names(pers) %>% 
  map_dfc(~ pw[,names(ok_samps[ok_samps >= ceiling(samp_num * .)])] %>% 
            ncol) %>% 
  gather( key = 'per', value = 'num')

# temporarily decides to choose 
per_thre <- 0.5

pw_fil <- pw[,names(ok_samps[ok_samps >= ceiling(samp_num * per_thre)])]

```

```{r}
# a look at the final pheno data
colnames(pheno_final)

cor.test(pheno_final$crp, pheno_final$IL_6, method = 'pearson')
```


```{r}
# remove everything already in this folder
system('rm ../data/maASlin2/pw/onlyFixed/*')


fit_data <- Maaslin2(
    pw, pheno_final, '../data/maASlin2/pw/onlyFixed', 
    normalization  = 'NONE', 
    transform = "LOG",  
    cores = 4, 
    plot_heatmap = F,
    min_abundance = 0.05,
    min_prevalence = 0.1,
    fixed_effects = c('Toxicity', 
                      'CR',
                      'Disease_Class',
                      'disease_sta',
                      'abx_expo', 
                      'ferritin',
                      'crp',
                      'IL_1_beta',
                      #'IL_10',
                      'IL_6',
                      'Prior_lines_of_therapy',
                      'sex',
                      'age',
                      'Antibiotics_4_weeks_post_CAR_T_cells'),
    standardize = FALSE,
    max_significance = 0.25)



```

```{r}
system('rm ../data/maASlin2/pw/both/*')

fit_data <- Maaslin2(
    pw, pheno_final, '../data/maASlin2/pw/both', 
    normalization  = 'NONE', 
    transform = "LOG",  
    cores = 4, 
    plot_heatmap = F,
    min_abundance = 0.05,
    min_prevalence = 0.1,
    fixed_effects = c('Toxicity', 
                      'CR',
                      'Disease_Class',
                      'disease_sta',
                      'abx_expo', 
                      'ferritin',
                      'crp',
                      'IL_1_beta',
                      #'IL_10',
                      'IL_6',
                      'Prior_lines_of_therapy',
                      'Antibiotics_4_weeks_post_CAR_T_cells'),
    random_effects = c('sex',
                      'age'),
    standardize = FALSE,
    max_significance = 0.25)
```
