==> scripts/29_abx_usage.Rmd <==
---
title: "ABX usage"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(tidyverse)
library(readxl)
library(lubridate)
```

```{r}
### Using the clinical data info from ctc.medications
ctc.medications.srx %>% 
  filter(MRN == '38009710' & drug_name == 'vancomycin') %>% 
  dplyr::select(MRN, drug_name, start_datetime, stop_datetime)


ctc.medications.srx %>% 
  filter(MRN == '35399272') %>% 
  dplyr::select(MRN, drug_name, start_datetime, stop_datetime) %>% 
  filter(str_detect(drug_name, 'cephalexin'))
  
```

```{r}
# the current shotgun data
sg <- read_csv('../data/cleaned_pheno/final_pheno.csv')

# the updated baseline sample collection date
bcd <- read_csv('/Volumes/castoricenter/CAR-Tcell_and_Microbiota/SampleDates12-12-2019.csv', 
                col_types = cols(.default = col_character())) %>% 
  # restore the MRN and pad necessary 0 in front of it
  mutate(mrn = str_pad(mrn, 8, side = 'left', pad = '0')) %>% 
  mutate(datecollection = mdy(datecollection)) %>% 
  filter(str_detect(window, 'Baseline')) %>% 
  filter(sampleid %in% sg$sampleid) 

# load the updated abx usage table
abx <- read_csv('../data/cleaned_pheno/copy12_12_abx.csv', col_types = cols(.default = col_character())) %>% 
  filter(MRN %in% bcd$mrn) %>% 
  dplyr::select(MRN, 
                Antibiotics, 
                start = `Start date`,
                end = `End Dates`) %>% 
  # there is one patient has None abx
  filter(!is.na(start)) %>% 
  mutate(start = mdy(start),
         end = mdy(end)) %>% 
  # join in the date of baseline sample collection
  left_join(bcd %>% 
              dplyr::select(mrn, datecollection), by = c('MRN'='mrn')) %>% 
  mutate(startday = start - datecollection,
         endday = end - datecollection)
```

```{r}
# create a table that has Y and N for abx usage 4 weeks prior to the baseline sample collection date
# only keep the MRN that have Y for abx 4 weeks prior to the baseline sample collection date (in the current TJ antibiotics table)
fil <- abx %>% 
  filter(startday <= 0 ) %>% 
  distinct(MRN) %>% 
  pull(MRN)


# the remaining MRNs that need to validate whether they have any abx
other <- setdiff(sg$MRN, fil)
```

## Start with Ying table and find out the abx medications

```{r}
# the below could only be run in the console but not a rmd trunk.
source('~/MSK/work/microbiome_db/SQL/scripts/get_data_from_query_OTU.R')
dbListTables(con) %>% data.frame() %>% filter(grepl("_ag",.)) 
abxtb <- get_data_from_query_OTU(0,'antibiotics_antimicrobial_ordername_map_ag') 
```

```{r}
# extract the unique abx names from the above table
abxu <- abxtb %>% 
  distinct(order_name) %>% 
  mutate(abx_name = str_to_lower(str_replace(order_name, ' .+$',''))) %>% 
  distinct(abx_name) %>% 
  pull(abx_name)
```

```{r}
# all medications of those 14 patients
othermed <- ctc.medications.srx %>% 
  filter(MRN %in% other) %>% 
  dplyr::select(MRN, med_name, start_date, stop_date) %>% 
  mutate(med_name = str_to_lower(str_replace(med_name, ' .+$',''))) %>% 
  filter(med_name %in% abxu) %>%  
  left_join(bcd %>% 
              dplyr::select(mrn, datecollection), by = c('MRN'='mrn')) %>% 
  mutate(startday = start_date - datecollection,
         endday = stop_date - datecollection) %>% 
  # selecting the records that are the proof of abx
  filter(startday >= -28 & startday <= 0) 

# the final MRN of the yes exposure to abx in the 4 w prior to baseline window
yexpo <- c(othermed %>% 
             distinct(MRN) %>% 
             pull(MRN), fil)
```

```{r}
# create a table to be joined with the final pheno table
abxout <- tibble(
  MRN = sg$MRN,
  abx_expo = if_else(MRN %in% yexpo, 'Y', 'N')
)

abxout %>% 
  write_csv('../data/cleaned_pheno/shotgun_cohort_abx_4w_baseline_expo_table.csv')
```
