==> scripts/31_univariate_lm_wilcox.Rmd <==
---
title: "Univariate model using Wilcox Rank Sum test"
author: "Anqi Dai"
date: "`r Sys.Date()`"
output:
  html_document:
    toc: true
    toc_float: true
    toc_depth: 3
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, message = F, warning = F)
```

```{r}
library(tidyverse)
library(kableExtra)
```
 

Univariate Wilcox test to see if there is significant pathways from the humann2 normalized data

## Load pheno and count data 

```{r}
final_pheno <- read_csv('../data/cleaned_pheno/final_pheno.csv')

pw <- read_tsv('../data/humann2_res/humann2_final_out/humann2_pathabundance_cpm_unstratified.tsv') %>% 
  dplyr::rename_all(funs(str_replace(., '_Abundance$',''))) %>% 
  dplyr::rename(pw = names(.)[1])  %>% 
  # remove the things after the : so that its easier to see on the plot(the pathway name)
  mutate(pw = str_replace(pw, ':.+$','')) 
```

##### Reorder the column and make sure the sample names in the pheno table align with the sample name in the count table

```{r}
pw <- bind_cols(pw[,1],pw[,final_pheno$sampleid])

all.equal(colnames(pw)[2:ncol(pw)], final_pheno$sampleid)

phe_tox <- final_pheno$Toxicity
phe_CR <- final_pheno$CR

pw <- pw %>% 
  column_to_rownames('pw') %>% 
  as.matrix() 
```

## Use different prevalence filtering threshold (percentage of zero count samples) and do the log2 transform and wilcox test

```{r}
num_zero_thre <- seq(0, 0.99, 0.05)

test_diff_zero_thre_filter <- function(fil_thre){
  
  num_zero_thre <- floor(ncol(pw) * fil_thre)
  
  pw_fil <- pw[rowSums(pw == 0) <= num_zero_thre, ]
  
  pw_log2 <- log2(pw_fil + 0.01)
  
  sig_tox <- apply(pw_log2, 1, function(r){
      res = wilcox.test(r ~ phe_tox, exact = F)
      return(pval = res$p.value)
    }) %>% 
      data.frame() %>% 
      rename(pval = names(.)[1]) %>% 
      rownames_to_column('pw') %>% 
      mutate(padj = p.adjust(pval, method = 'BH')) %>% 
      arrange(desc(padj)) %>% 
      filter(padj < 0.25) %>% 
      nrow
  
  sig_CR <- apply(pw_log2, 1, function(r){
      res = wilcox.test(r ~ phe_CR, exact = F)
      return(pval = res$p.value)
    }) %>% 
      data.frame() %>% 
      rename(pval = names(.)[1]) %>% 
      rownames_to_column('pw') %>% 
      mutate(padj = p.adjust(pval, method = 'BH')) %>% 
      arrange(desc(padj)) %>% 
      filter(padj < 0.25) %>% 
      nrow
  
  return(list(sig_tox_num = sig_tox,
         sig_CR_num = sig_CR,
         pw_num = nrow(pw_log2)))
}


res <- list()
for (i in 1:length(num_zero_thre)){
  res[[i]] = test_diff_zero_thre_filter(num_zero_thre[i])
}

final <- bind_rows(res) %>% 
  mutate(zero_perc = num_zero_thre) %>% 
  select(zero_perc, pw_num, sig_tox_num, sig_CR_num)


final %>% 
  kable(caption = 'Number of significant results (FDR < 0.25) in different zero percentage threshold') %>% 
  kable_styling(full_width = F, position = 'left')
```

```{r}
final
```
