==> scripts/39_new_MSK_shotgun.Rmd <==
---
title: "New MSK samples"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
library(tidyverse)
```

```{r}
# check the sequencing files returned, how the fid is matched to sampleid, is there any duplicates etc
proj <- c('Project_11454') # there are three failed QC so this is the correct number 

folder <- '/Volumes/vandenBrinkLab/deep_sequencing'

files <- folder %>% 
  map_dfr(function(f){
    proj %>% 
      set_names(proj) %>% 
      imap_dfr( ~ Sys.glob(str_glue('{f}/{..1}/*/*')) %>% 
             as.data.frame %>% 
              mutate(..2)) %>% 
      rename(directory = names(.)[1],
             projectID = names(.)[2])
  }) %>% 
  mutate(sample_folder = str_extract(directory, 'Sample_.+$'),
         fid = str_replace(sample_folder, 'Sample_',''),
         fid = str_replace(fid, '_IGO_.+$','')) %>% 
  mutate(sampleid = fid) # the fid looks clean should be the sampleid 

files %>% 
  distinct(sampleid) %>% 
  nrow

files %>% 
  write_csv('../data/shotgun_input/Project_11454_shotgun_samples.csv')

# although 20 returned but few can be used, the non-useable ones are from ineligible pts
```

```{r}
# check the sample patient pheno table organized by John
# the first sheet is the eligible samples
# the below table has the most updated info organized by John
stb <- readxl::read_excel('/Volumes/castoricenter/CAR-Tcell_and_Microbiota/MSK/samples/sample_table_msk.xlsx', sheet = 1) 
ineligible <- readxl::read_excel('/Volumes/castoricenter/CAR-Tcell_and_Microbiota/MSK/samples/sample_table_msk.xlsx', sheet = 2)

pheno <- read_csv('../data/cleaned_pheno/final_pheno.csv') # the 33 used last year
intersect(pheno$sampleid, ineligible$sampleid)  # 6 has become ineligible 

setdiff(current_Baseline$sampleid, pheno$sampleid)

missing <- c("2460A", "2563A", "1918C" ,"315"  )
```

## all MSK eligible baseline shotgun samples

2020-12-7   32 total for eligible samples in the baseline time point from MSK patients

```{r}
current_Baseline <- stb %>% 
  inner_join(shotgun_lookup_ad) %>% 
  filter(str_detect(time_group, '^Baseline')) %>% 
  distinct(sampleid, .keep_all = T)

ptb <- readxl::read_excel('/Volumes/castoricenter/CAR-Tcell_and_Microbiota/MSK/patients/msk_essential_data_v4.xlsx') %>% 
  mutate(mrn = as.numeric(mrn))

pheno_msk <- current_Baseline %>% 
  select(mrn:sampleid, fid, directory) %>% 
  left_join(ptb)


pheno_msk %>% 
  write_csv('../data/shotgun_pheno/msk_all_pheno.csv')

pheno_msk %>% 
  filter(sampleid %in% missing)
```
