==> scripts/40_combine_all_samp_pheno.Rmd <==
---
title: "combine all centers pheno"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
library(tidyverse)
```

This is the script that sorts out the pheno table for shotgun data.

```{r}
# all centers pheno file 
# to get the baseline sampleid 
I <- read_csv('../data/shotgun_pheno/italy_all_earlier.csv') %>% 
  filter(timepoint < 0) %>% 
  ungroup() %>% 
  select(fid)

I2 <- read_csv('../data/shotgun_pheno/italy_all_later.csv') %>% 
  select(fid)

P <- read_csv('../data/shotgun_pheno/upenn_all_earlier.csv') %>% select(fid)

P2 <- read_csv('../data/shotgun_pheno/upenn_all_later.csv') %>% 
  select(fid)

M <- read_csv('../data/shotgun_pheno/msk_all_pheno.csv') %>% select(fid)

all_baseline <- bind_rows(I,I2, M, P, P2)

all_baseline %>% 
  write_csv('../data/shotgun_input/samples_cart_all_msk_patial_italy_upenn.csv', col_names = F)
```

## the pheno label for the samples

```{r}
# trying to get back the timepoint, sampleid, oligoid
# starting from the patient and their earliest sample better have pre in their raw sampleid

I2 <- read_csv('../data/shotgun_pheno/italy_all_later.csv')

isamp <- readxl::read_excel('/Volumes/castoricenter/CAR-Tcell_and_Microbiota/Italy/samples/opbg_sampleskey_with_oligoid_v2.xlsx', sheet = 1) %>% 
  filter(pt_id %in% I2$pid) %>% 
  filter(timepoint < 0) %>% 
  distinct(sampleid, .keep_all = T) %>% 
  full_join(I2 %>% select(pt_id = pid, fid))

P <- read_csv('../data/shotgun_pheno/upenn_all_earlier.csv') 

P2 <- read_csv('../data/shotgun_pheno/upenn_all_later.csv') 

M <- read_csv('../data/shotgun_pheno/msk_all_pheno.csv')
```


```{r}
I1 <- read_csv('../data/shotgun_pheno/italy_all_earlier.csv') %>% 
  filter(timepoint < 0) %>% 
  ungroup() 

# the pheno lable for the I2!
iptb <- readxl::read_excel('/Volumes/castoricenter/CAR-Tcell_and_Microbiota/Italy/patients/italy_essential_data_v4.xlsx', sheet = 1)
I2_pheno <- isamp %>% 
  left_join(iptb) 

I_phe <- bind_rows(
  I1 %>% 
    select(fid, pt_id = pid, timepoint:last_followup), 
  I2_pheno %>% 
    select(fid, pt_id, timepoint:sampleid, oligoid:last_followup)
) %>% 
  mutate(center = 'I')
```

```{r}
# upenn
P_ptb <- readxl::read_excel('/Volumes/castoricenter/CAR-Tcell_and_Microbiota/U Penn/patients/upenn_essential_data_v3.xlsx', sheet = 1)

P_ptb_ineligible <- readxl::read_excel('/Volumes/castoricenter/CAR-Tcell_and_Microbiota/U Penn/patients/upenn_essential_data_v3.xlsx', sheet = 2)

P_ <- bind_rows(
  P %>% 
    mutate(pt_id = str_replace(fid, 'MS_UPENN_','')) %>% select(fid, pt_id) %>% mutate(pt_id = as.numeric(pt_id)),
  P2 %>% 
    mutate(pt_id = as.numeric(fid)) %>% select(fid, pt_id)
) %>% 
  filter(! pt_id %in% P_ptb_ineligible$pt_id)

P_phe <- P_ %>% 
  left_join(P_ptb)  %>% 
  mutate(center = 'P') 


psamp <- readxl::read_excel('/Volumes/castoricenter/CAR-Tcell_and_Microbiota/U Penn/samples/sample_mapping_UPenn.xlsx', sheet = 1) %>% 
  mutate(pt_id = pt_id-8000) %>% 
  filter(pt_id %in% P_phe$pt_id) %>% 
  full_join(P_phe) %>% 
  select(colnames(I_phe))
```


```{r}
# MSK
msamp <- readxl::read_excel('/Volumes/castoricenter/CAR-Tcell_and_Microbiota/MSK/samples/sample_table_msk.xlsx', sheet = 1)
M_phe <- M %>% 
  mutate(timepoint = NA,
         oligoid = NA) %>% 
  mutate(center = 'M',
         pt_id = mrn) %>% 
  select(colnames(I_phe))
```


```{r}
# combined 

all <- bind_rows(
  M_phe %>% mutate(pt_id = as.character(pt_id)), 
  psamp %>% 
    mutate(pt_id = as.character(pt_id)), 
  I_phe 
)

# save a copy to the server for John
all %>% 
  write_csv('/Volumes/castoricenter/CAR-Tcell_and_Microbiota/shotgun_results/pheno_data_comprehensive.csv')

# save one for my own
all %>% 
  write_csv('../data/shotgun_pheno/final_comprehensive.csv')

all <- read_csv('../data/shotgun_pheno/final_comprehensive.csv')
```

## tally of the table

```{r}
# make proper columns factors as they should be 
should_be_factors <- all %>% 
  select_if(is.character) %>% 
  select(-fid, -pt_id) %>% 
  colnames()

ALL <- all %>% 
  mutate_at(should_be_factors, as.factor)


ALL %>% summary
```


```{r}
outcome <- c('cr_d100','toxicity','icans','crs')

ALL %>% 
  count(center, cr_d100)

ALL %>% 
  count(center, toxicity)

ALL %>% 
  count(center, icans)

ALL %>% 
  count(center, crs)

tb <- ALL %>% 
  select(center, outcome) 
```

```{r}
tally <- ALL %>% 
  select(center, outcome) %>% 
  mutate_if(is.factor, as.character) %>% 
  gather('metric', 'value', cr_d100 : crs) 
  

individual <- tally %>% 
  count(center, metric, value) %>% 
  filter(value != 'not_assessed') %>% 
  spread('metric', 'n', fill = 0)

pull_ <- tally %>% 
  count(metric, value)  %>% 
  filter(value != 'not_assessed') %>% 
  mutate(center = 'pull') %>% 
  spread('metric', 'n', fill = 0) %>% 
  select(colnames(individual))

tally_all <- bind_rows(
  individual, pull_
)
  
```


```{r}
library(kableExtra)

kbl(tally_all, caption = "") %>%
  kable_minimal( full_width = F) %>%
  pack_rows("Group Italy", 1, 2) %>%
  pack_rows("Group MSK", 3, 4) %>% 
  pack_rows("Group UPENN", 5, 6)  %>% 
  pack_rows("Group ALL", 7, 8)


```
