==> scripts/43_comparisons_3.Rmd <==
---
title: "Comparisons in three center"
output: html_document
--- 

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
library(tidyverse)
```

SUBSTITUTE WITH HUMANN3 RESULTS WHEN I CAN 

```{r}
# UPDATE THIS TABLE BABY
cur <- read_csv('../data/shotgun_pheno/final_comprehensive.csv') %>% 
  select(fid, center, cr_d100, toxicity, icans, crs) %>% 
  filter(center != 'I') %>% # manually update the label for the "not_assessed"
  mutate(cr_d100 = case_when(
    fid == '2460A' ~ 'no',
    fid == '2293A' ~ 'no',
    fid == '2563A' ~ 'no',
    TRUE ~ cr_d100
  ))


# Removing the patients that the data we can't use
# upated 2021-4-12
now <- read_csv('/Volumes/castoricenter/CAR-Tcell_and_Microbiota/MSK/samples/eligible_sampleslist_JS.csv') %>% 
  filter(eligible == 'yes' & shotgun == 'yes') %>% 
  pull(pt_id)

final <- read_csv('../data/shotgun_pheno/final_comprehensive.csv') %>% 
  filter(center != 'I') %>% 
  mutate(pt_id = as.numeric(pt_id)) %>% 
  mutate(pt_id = if_else(center == 'P', pt_id + 8000, pt_id)) %>% 
  filter(pt_id %in% now) # keep to the most updated list 

final %>% write_csv('../data/shotgun_pheno/final_comprehensive_UPDATED.csv')

# the final simple list with most up to date pheno label
simple <- cur %>% 
  filter(fid %in% final$fid)

simple %>% write_csv('../data/shotgun_pheno/final_comprehensive_UPDATED_simple.csv')

simple <- read_csv('../data/shotgun_pheno/final_comprehensive_UPDATED_simple.csv')

pheno <- simple %>% 
  gather('pheno', 'value', cr_d100:crs) 
```


```{r}


sub_phenos <- pheno %>% 
  split(., list(.$center, .$pheno)) %>% 
  purrr::imap(~ filter(.data = ., value != 'not_assessed'))


```

```{r}
# the pheno for the different center and tox/CR_d30
sub_phenos %>% 
  imap(function(.x, .y){
    select(.data = .x, value) %>% 
      t() %>% 
      write.table(str_glue('../data/shotgun_output/humann3/{.y}.txt'), sep = '\t', quote = F, row.names = T, col.names = F)
  })
```

```{r}
# the pathway counts table
full <- read_tsv('../data/shotgun_output/humann3_tsv/humann3_pathabundance_cpm_joined_unstratified.tsv')  %>% 
  rename_all(funs(str_replace(., '_humann3$','')))

pcts <- sub_phenos %>% 
  map(~ pull(.data = ., fid) ) %>% 
  imap(~ full %>% select(`# Pathway`, matches(.x)) %>% 
         write_tsv(str_glue('../data/shotgun_output/humann3/{.y}_pcts.tsv'))) 
```

## the genefamilies regrouping

```{bash}
# regroup the whole gene family table here
humann_regroup_table --input humann3_genefamilies_CART.tsv --groups uniref90_ko --output humann3_KO_CART.tsv
```

```{r}
fullko <- read_tsv('../data/shotgun_output/humann3_tsv/humann3_KO_CART.tsv')  %>% 
  rename_all(funs(str_replace(., '^CART_','')))  %>% 
  rename_all(funs(str_replace(., '_humann3$','')))

kocts <- sub_phenos %>% 
  map(~ pull(.data = ., fid) ) %>% 
  imap(~ fullko %>% select(`# Gene Family`, matches(.x)) %>% 
         write_tsv(str_glue('../data/shotgun_output/humann3/{.y}_kocts.tsv'))) 
```

```{r}
# the lefse for the ko only
fns <- list.files('../data/shotgun_output/humann3/', pattern = 'lefse_ready_kocts.tsv$')

cmds <- tibble(
  fns = fns
) %>% 
  mutate(format_cmd = str_glue('format_input.py {fns}  {fns}.in -c 1 -u 2 -o 1000000')) %>% 
  mutate(run_cmd = str_glue('run_lefse.py {fns}.in  {fns}.res')) %>% 
  mutate(plot_cmd = str_glue('plot_res.py {fns}.res {fns}.png  --feature_font_size 4 --width 10 --dpi 300 --title {fns}')) %>% 
  select(-fns) %>% 
  gather() %>% 
  select(value) %>% 
  write_csv('../data/shotgun_output/humann3/lefse_run_cmd_KO.sh', col_names = F)
```

## run both of them with all of the samples pulled together

```{r}
all_sub_pheno <- pheno %>% 
  split(., list(.$pheno)) %>% 
  purrr::imap(~ filter(.data = ., value != 'not_assessed'))

all_sub_pheno %>% 
  imap(function(.x, .y){
    select(.data = .x, value) %>% 
      t() %>% 
      write.table(str_glue('../data/shotgun_output/humann3/pull_{.y}.txt'), sep = '\t', quote = F, row.names = T, col.names = F)
  })
 

all_pcts <- all_sub_pheno %>% 
  map(~ pull(.data = ., fid) ) %>% 
  imap(~ full %>% select(`# Pathway`, matches(.x)) %>% 
         write_tsv(str_glue('../data/shotgun_output/humann3/pull_{.y}_pcts.tsv')))
```

```{r}
# add a filtering step here after the normalization in the snakemake 
# 50 and 25% 
pcts <- read_tsv('../data/shotgun_output/humann3/humann_final_out/pull_cr_d100_pcts_cpm_unstratified.tsv') %>% 
  gather('sampleid', 'cpm', names(.)[2]:names(.)[ncol(.)]) %>% 
  rename(pw = `# Pathway`)

keeppw <- pcts %>% 
  filter(cpm > 50) %>%
  ungroup() %>% 
  count(pw) %>% 
  filter(n > floor(nrow(simple) * 0.25)) %>% 
  pull(pw)

cts_fil <- pcts %>% 
  filter(pw %in% keeppw) %>% 
  spread('sampleid', 'cpm', fill = 0)

cts_fil %>% 
  write_tsv('../data/shotgun_output/humann3/humann_final_out/pull_cr_d100_pcts_cpm_unstratified_fil.tsv')
```

```{r}
# tox
pcts <- read_tsv('../data/shotgun_output/humann3/humann_final_out/pull_toxicity_pcts_cpm_unstratified.tsv') %>% 
  gather('sampleid', 'cpm', names(.)[2]:names(.)[ncol(.)]) %>% 
  rename(pw = `# Pathway`)

keeppw <- pcts %>% 
  filter(cpm > 50) %>%
  ungroup() %>% 
  count(pw) %>% 
  filter(n > floor(nrow(simple) * 0.25)) %>% 
  pull(pw)

cts_fil <- pcts %>% 
  filter(pw %in% keeppw) %>% 
  spread('sampleid', 'cpm', fill = 0)

cts_fil %>% 
  write_tsv('../data/shotgun_output/humann3/humann_final_out/pull_toxicity_pcts_cpm_unstratified_fil.tsv')

# and then bind with snakemake
```


```{r}
# KO
all_kocts <- all_sub_pheno %>% 
  map(~ pull(.data = ., fid) ) %>%  
  imap(~ fullko %>% select(`# Gene Family`, matches(.x)) %>% 
         write_tsv(str_glue('../data/shotgun_output/humann3/pull_{.y}_kocts.tsv')))
```


```{r}
all_gfcts <- all_sub_pheno %>% 
  map(~ pull(.data = ., fid) ) %>% 
  imap(~ fullgenefam %>% select(`# Gene Family`, matches(.x)) %>% 
         write_tsv(str_glue('../data/shotgun_output/humann3/pull_{.y}_gfcts.tsv'))) 
```

## Need to go to snakemake for some steps related to humann before the lefse



## generate command to run lefse in py2 environment

cuz can't get them thru snakemake
 
```{r}
# PATHWAY
# I already normalized the pcts so don't need to normalize again here 
fns <- list.files('../data/shotgun_output/humann3/', pattern = 'lefse_ready_pcts.tsv$')
    
cmds <- tibble(
  fns = fns
) %>% 
  mutate(format_cmd = str_glue('format_input.py {fns}  {fns}.in -c 1 -u 2')) %>% 
  mutate(run_cmd = str_glue('run_lefse.py {fns}.in  {fns}.res')) %>% 
  mutate(plot_cmd = str_glue('plot_res.py {fns}.res {fns}.pdf --format pdf  --feature_font_size 4 --width 10 --dpi 300 --title {fns}')) %>% 
  select(-fns) %>% 
  gather() %>% 
  select(value) %>% 
  write_csv('../data/shotgun_output/humann3/lefse_run_cmd.sh', col_names = F)
```


## The taxa relative abundance from the shotgun data using metaphlan

```{r}
taxa <- read_tsv('../data/shotgun_output/metaphlan/merged_abundance_table.txt')  %>% 
  rename_all(funs(str_replace(., '_profile$','')))

keeptaxa <- taxa %>% 
  filter(cpm > 50) %>%
  ungroup() %>% 
  count(pw) %>% 
  filter(n > floor(nrow(simple) * 0.25)) %>% 
  pull(pw)

cts_fil <- pcts %>% 
  filter(pw %in% keeppw) %>% 
  spread('sampleid', 'cpm', fill = 0)

tcts <- sub_phenos %>% 
  map(~ pull(.data = ., fid) ) %>% 
  imap(~ taxa %>% select(clade_name, matches(.x)) %>% 
         write_tsv(str_glue('../data/shotgun_output/humann3/{.y}_tcts.tsv'))) 

all_tcts <- all_sub_pheno %>% 
  map(~ pull(.data = ., fid) ) %>% 
  imap(~ taxa %>% select(clade_name, matches(.x)) %>% 
         write_tsv(str_glue('../data/shotgun_output/humann3/pull_{.y}_tcts.tsv')))

# then just need to use the snakemake rule to cat the two
```

```{r}
# run lefse
fns <- list.files('../data/shotgun_output/humann3/', pattern = 'lefse_ready_tcts.tsv$')

cmds <- tibble(
  fns = fns
) %>% 
  mutate(format_cmd = str_glue('format_input.py {fns}  {fns}.in -c 1 -u 2 -o 1000000')) %>% 
  mutate(run_cmd = str_glue('run_lefse.py {fns}.in  {fns}.res')) %>% 
  mutate(plot_cmd = str_glue('plot_res.py {fns}.res {fns}.pdf --format pdf --feature_font_size 4 --width 10 --dpi 300 --title {fns}')) %>% 
  mutate(clado_cmd = str_glue('plot_cladogram.py {fns}.res {fns}_clado.pdf  --label_font_size 4 --dpi 300 --format pdf --title {fns}')) %>% 
  select(-fns) %>% 
  gather() %>% 
  select(value) %>% 
  write_csv('../data/shotgun_output/humann3/lefse_run_cmd_taxa.sh', col_names = F)
```


## PCOA to see if batch effects

```{r}
# very quicly done some pcoa
cts <- read_tsv('../data/shotgun_output/humann2/humann_final_out/pull_toxicity_pcts_cpm_unstratified.tsv') %>% 
  gather('sampleid', 'value', names(.)[2]:names(.)[ncol(.)]) %>% 
  rename(pathway =  names(.)[1]) %>% 
  spread('pathway', 'value') %>% 
  column_to_rownames('sampleid')

dist_bl <- vegan::vegdist(cts, method = 'bray')

bl_bc <- cmdscale(dist_bl, k = 2) 
pheno <- read_csv('../data/shotgun_pheno/final_simplified.csv') %>% 
  mutate(sampleid = fid)

bl_bc %>% 
    as.data.frame() %>% 
    rownames_to_column('sampleid') %>% 
    inner_join(pheno, by = 'sampleid') %>% 
    ggpubr::ggscatter(x = 'V1', y = 'V2', color =  'center',  shape = 'toxicity') 
```
```{r}
empty_samp <- read_tsv('../data/shotgun_output/humann2/humann_final_out/pull_toxicity_kocts_cpm_unstratified.tsv') %>% 
  gather('sampleid', 'value', names(.)[2]:names(.)[ncol(.)]) %>% 
  group_by(sampleid) %>% 
  summarise(Sum = sum(value)) %>% 
  filter(Sum == 0) %>% 
  pull(sampleid)


cts <- read_tsv('../data/shotgun_output/humann2/humann_final_out/pull_toxicity_kocts_cpm_unstratified.tsv') %>% 
  gather('sampleid', 'value', names(.)[2]:names(.)[ncol(.)]) %>% 
  rename(KO =  names(.)[1]) %>% 
  filter(!sampleid %in% empty_samp) %>% 
  spread('KO', 'value') %>% 
  column_to_rownames('sampleid')


  

dist_bl <- vegan::vegdist(cts, method = 'bray')

bl_bc <- cmdscale(dist_bl, k = 2) 
pheno <- read_csv('../data/shotgun_pheno/final_simplified.csv') %>% 
  mutate(sampleid = fid)

bl_bc %>% 
    as.data.frame() %>% 
    rownames_to_column('sampleid') %>% 
    inner_join(pheno, by = 'sampleid') %>% 
    ggpubr::ggscatter(x = 'V1', y = 'V2', color =  'center',  shape = 'toxicity') 
```
