==> scripts/44_ggplot_venn_lefse.Rmd <==
---
title: "ggplot and the venn diagram for the lefse bars !"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
library(tidyverse)
library(ggvenn)
```

Create a function of ggplot viewing the lefse res files and also venn to see the overlap and whatnot of the different comparisons


## collect all of the res files

```{r}
res <- list.files('../data/shotgun_output/humann3/', pattern = 'lefse_ready_.+cts.tsv.res$', full.names = T)

res_all <- res %>% 
  set_names(res) %>% 
  map(~ read_tsv(., col_names = c('feature','xx','direction','score','pval')) %>% 
        filter(!is.na(score))) %>% 
  keep(~ nrow(.) > 0) %>% 
  bind_rows(.id = 'res') %>% 
  mutate(tbl = if_else(str_detect(res, 'pcts'), 'pathway', 'taxa')) %>% 
  mutate(res = str_replace(res, '^.+//',''),
         res = str_replace(res, '_lefse.+$',''))  %>% 
  rename(grp = res) %>% 
  mutate(grp = str_replace(grp, 'pull_','pull\\.')) %>% 
  separate(grp, into = c('cohort','outcome'), sep = '\\.') 

res_all %>% 
  ungroup() %>% 
  filter(cohort == 'pull') %>% 
  write_csv('../data/shotgun_output/pull_pathway_lefse.csv')

res_all_list <- res_all %>% 
  split(.$tbl) %>% 
  map(~split(., list(.$cohort, .$outcome )) %>% 
  map(~ pull(.data = ., feature)) %>% 
  compact())

res_all %>% 
  write_csv('../data/shotgun_output/lefse_sig_results_taxa_and_pw.csv')

```


```{r}
vars <- res_all %>% 
  distinct(outcome) %>% 
  pull(outcome)


tbs <- names(res_all_list)

tbs %>% 
  set_names(tbs) %>% 
  map(function(tb){
    vars %>% 
      set_names(vars) %>% 
      map(function(var){
        res_all_list %>% 
          pluck(tb) %>% 
          keep(.p = str_detect(names(.), var))
  } ) %>% 
      imap(~ ggvenn(.x) +
             labs(title = str_glue('{tb}')) +
          ggsave(str_glue('../figs/{tb}_venn_{.y}.pdf'), width = 7, height = 4))
  })
```

```{r}
# for the KO table
library(KEGGREST)
res_all_list %>% pluck('KO') %>% 
  keep(str_detect(names(.), 'toxicity')) %>% 
  reduce(intersect) %>% 
  map(~ list(KOID = keggGet(.)[[1]]$ENTRY,
             KO_name = keggGet(.)[[1]]$NAME,
             KO_def = keggGet(.)[[1]]$DEFINITION),
             KO_pathway = keggGet(.)[[1]]$PATHWAY) %>%
  bind_rows()
```


## a PCOA to see how the data

```{r}
cts <- read_tsv('../data/shotgun_output/humann3/humann_final_out/pull_toxicity_kocts_cpm_unstratified.tsv') %>% 
  gather('sampleid', 'value', names(.)[2]:names(.)[ncol(.)]) %>% 
  rename(pathway =  names(.)[1]) %>% 
  spread('pathway', 'value') %>% 
  column_to_rownames('sampleid')
dist_bl <- vegan::vegdist(cts, method = 'bray')
bl_bc <- cmdscale(dist_bl, k = 2) 
pheno <- read_csv('../data/shotgun_pheno/final_comprehensive.csv') %>% 
  mutate(sampleid = fid)
bl_bc %>% 
    as.data.frame() %>% 
    rownames_to_column('sampleid') %>% 
    inner_join(pheno, by = 'sampleid') %>% 
    ggpubr::ggscatter(x = 'V1', y = 'V2', color =  'center',  shape = 'toxicity') 
```

## output a pretty ggplot version of the lefse bar plot

```{r}
make <- res_all %>% 
  filter(cohort == 'pull' & tbl == 'pathway') %>% 
  mutate(score = if_else(direction == 'no', -score, score)) %>% 
  mutate(feature = str_replace_all(feature, '_',' ')) 
  #mutate(feature = str_wrap(feature, width = 55, indent = 0, exdent = 0))


all_title_fs <- 20
axis_text_fs <- 16
make %>% 
  split(.$outcome) %>% 
  imap(~   ggplot(data = .x, aes(x = reorder(feature, score), y = score, fill = direction)) +
            geom_col(position = position_dodge2(width = 0.9, preserve = "single")) +
            coord_flip() +
            #scale_color_manual(values = c('#0099B4', '#AD002A'))  +
            #scale_fill_manual(values = c('#0099B4', '#AD002A')) +
            theme_classic() +
            theme(axis.title.y  = element_blank(), 
                  axis.title.x = element_text(size=axis_text_fs),
                  plot.title = element_text(size=all_title_fs),
                  axis.text.x = element_text(size=axis_text_fs),
                  legend.position="bottom") +
         scale_fill_manual(values = c('#0099B4','#AD002A')) +
            labs(title = str_glue('Response to {.y}') ,
                 y = 'Score')+ 
            ggsave(str_glue('../data/shotgun_output/humann3/pull_{.y}_pathway.pdf'), device = 'pdf', width = 12, height = 20))
            #ggsave(str_glue('/Volumes/castoricenter/CAR-Tcell_and_Microbiota/shotgun_results/NEW_humann3_version/pathway/pull_{.y}_pathway.pdf'), device = 'pdf', width = 12))
            
make %>% 
  filter(outcome == 'cr_d100') %>% 
  split(.$outcome) %>% 
  imap(~   ggplot(data = .x, aes(x = reorder(feature, score), y = score, fill = direction)) +
            geom_col(position = position_dodge2(width = 0.9, preserve = "single")) +
            coord_flip() +
            #scale_color_manual(values = c('#0099B4', '#AD002A'))  +
            #scale_fill_manual(values = c('#0099B4', '#AD002A')) +
            theme_classic() +
            theme(axis.title.y  = element_blank(), 
                  axis.title.x = element_text(size=axis_text_fs),
                  plot.title = element_text(size=all_title_fs),
                  axis.text.x = element_text(size=axis_text_fs),
                  legend.position="bottom") +
         scale_fill_manual(values = c('#ED0000','#00468B')) +
            labs(title = str_glue('Response to {.y}') ,
                 y = 'Score')+ 
            ggsave(str_glue('../data/shotgun_output/humann3/pull_{.y}_pathway.pdf'), device = 'pdf', width = 8))

# for above this is for the Italy excluded version 
```
