==> scripts/45_16s_sorting_out.Rmd <==
---
title: "Sorting out the 16s"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
library(tidyverse)
library(readxl)
library(ggpubr)
```

This script will be cleaning and organizing the 16s samples 

```{r}
# upenn is all baseline samples 
ustb <- read_excel('/Volumes/castoricenter/CAR-Tcell_and_Microbiota/U Penn/samples/sample_mapping_UPenn.xlsx') 

uptb <- read_excel('/Volumes/castoricenter/CAR-Tcell_and_Microbiota/U Penn/patients/upenn_essential_data_v3.xlsx', sheet = 3) %>% 
  filter(sequenced == 'yes')
```

```{r}
# MSK contains baseline and following time points
mstb_ineli <- read_excel('/Volumes/castoricenter/CAR-Tcell_and_Microbiota/MSK/samples/sample_table_msk.xlsx', sheet = 2)
mstb <- read_excel('/Volumes/castoricenter/CAR-Tcell_and_Microbiota/MSK/samples/sample_table_msk.xlsx', sheet = 1) %>% 
  filter(! sampleid %in% mstb_ineli$sampleid)
 
mptb_ineli <-  read_excel('/Volumes/castoricenter/CAR-Tcell_and_Microbiota/MSK/patients/msk_essential_data_v4.xlsx', sheet = 2)
mptb <-  read_excel('/Volumes/castoricenter/CAR-Tcell_and_Microbiota/MSK/patients/msk_essential_data_v4.xlsx', sheet = 1) %>% 
  filter(! pt_id %in% mptb_ineli$pt_id)
  
```

## baseline samples only meta

```{r}
m_meta_b <- mstb %>% 
  filter(str_detect(time_group, '^Baseline')) %>% 
  inner_join(mptb, by = c('mrn' = 'pt_id'))


u_meta_b <- ustb %>% 
  inner_join(uptb, by = "pt_id")

colnames(u_meta_b)
```


```{r}
colnames(m_meta_b)
```

## Look at the tally of the phenotypes 

```{r}
m_meta_b %>% 
  count(cr_d30)

m_meta_b %>% 
  count(cr_d100)

m_meta_b %>% 
  count(toxicity)
```
```{r}
u_meta_b %>% 
  count(cr_d30)

u_meta_b %>% 
  count(cr_d100)

u_meta_b %>% 
  count(toxicity)
```
## merge the two pheno tables and then write out

```{r}
m <- m_meta_b %>% 
  select(-pt_id_old) %>% 
  rename(pt_id = mrn) %>% 
  select(-last_followup) %>% 
  mutate(center = 'M') %>% 
  select(pt_id:vital_status, center) # there are several more columns in the msk table now 

p <- u_meta_b %>%   
  mutate(time_group = 'Baseline') %>% 
  select(-sampleid) %>% 
  rename(sampleid = oligoid) %>% 
  select(-last_followup) %>% 
  mutate(center = 'P')  %>% 
  select(colnames(m))  
  
combined <- bind_rows(m, p)
```

## also get the alpha diversity here 

```{r}
library(vdbR)
connect_database('~/dbConfig.txt')
get_table_from_database('asv_alpha_diversity_ag')

combined_join <- combined %>% 
  inner_join(asv_alpha_diversity_ag %>% 
               select(sampleid, simpson_reciprocal))

combined_join %>% 
  write_csv('../data/amplicon/stool/combined_2_meta.csv')

combined_join %>% 
  write_csv('/Volumes/castoricenter/CAR-Tcell_and_Microbiota/Angel_amplicon/16s_data_pheno_comprehensive.csv')
```

```{r}
# check something about abx
m <- readxl::read_excel('/Volumes/castoricenter/CAR-Tcell_and_Microbiota/MSK/antibiotics/sean_table/abx_table_msk_sean.xlsx') %>% 
  select(pt_id, anaerobe_targeting)
meta <- read_csv('../data/amplicon/stool/combined_2_meta_expanded.csv') %>% 
  left_join(m) %>% 
  mutate(anaerobe_targeting = if_else(is.na(anaerobe_targeting), 'no', anaerobe_targeting))

m <- meta %>% 
  filter(center == 'M')

m %>% 
  count(anaerobe_targeting, cr_d100) %>% 
  ggbarplot(x = 'anaerobe_targeting', y= 'n', color = 'cr_d100', fill = 'cr_d100', label = TRUE)
```
