==> scripts/48_16s_beta_diversity.Rmd <==
---
title: "16s beta diversity"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
library(vegan)
library(tidyverse)
set.seed(123)
```

Do a PcoA and do the anosim test

**Do M, P and pull separately, and then do for each of the 4 things CR_d100, Toxicity, ICANS and CRS**

```{r}
genus <- read_csv('../data/amplicon/stool/counts_genus_combined.csv')
meta <- read_csv('../data/amplicon/stool/combined_2_meta.csv') 
pheno <- meta %>% 
  select(sampleid, cr_d100, toxicity, crs, icans)

nsamp <- genus %>% distinct(sampleid) %>% nrow

# do some filtering 10% and 0.01%
keep <- genus %>% 
  filter(relab > 0.0001) %>% 
  count(genus) %>% 
  filter(n > floor(nsamp * 0.2)) %>% 
  filter(!is.na(genus)) %>% 
  pull(genus) 

# find the dominating genus in each sample
domi <- genus %>% 
  arrange(sampleid, -relab) %>% 
  group_by(sampleid) %>% 
  slice(1) %>% 
  rename(domgenus = genus) %>% 
  select(sampleid, domgenus)
  

dat <- genus %>% 
  filter(genus %in% keep) %>% 
  spread(key = 'genus', value = 'relab') %>% 
  full_join(pheno) %>% 
  full_join(domi)

```
 
```{r}
# a function to do the combinations of the anosim test
pheno_ <- colnames(pheno)[2:5] 
n <- length(pheno_)

combi <- data_frame(
  g1 = c(rep('M',n), rep('P',n), rep('M',n)),
  g2 = c(rep('M',n), rep('P',n), rep('P',n)),
  label = rep(pheno_, 3)
)
   
do_anosim_each <-  function(g1_, g2_, label_){

    df = dat %>% 
      filter(center %in% c(g1_, g2_)) %>% 
      filter(get(label_) %in% c('yes','no')) 
    cts = df %>% 
      select(sampleid, all_of(keep)) %>% 
      column_to_rownames('sampleid')
    phe = df %>% 
      select(matches(label_)) %>% pull(get(label_)) 
    phedf = df %>% 
      select(sampleid, matches(label_))

    # anosim run
    ret = anosim(cts, phe, permutations = 999, distance = "bray", strata = NULL, parallel = 4)

    # plot the pcoa
    dist_ <- vegdist(cts, method = 'bray')
    bc <- cmdscale(dist_, k = 2)
    bc %>%
      as.data.frame() %>%
      rownames_to_column('sampleid') %>%
      full_join(phedf, by = 'sampleid') %>%
      ggplot(aes_string(x = "V1", y = "V2", color =  label_)) + geom_point() +
      labs(title = str_glue('PCoA {g1_}_{g2_}__{label_}\npval:{ret$signif} statistic:{round(ret$statistic, 2)}')) +
      theme_void() +
      ggsave(str_glue('../figs/beta_{g1_}_{g2_}__{label_}.pdf'))

  return(list(pval = ret$signif,
              statistic = ret$statistic))
}
 
all_anosim <- pmap(combi, function(g1, g2, label){
    do_anosim_each(g1, g2, label)
  }) %>% 
  set_names(str_glue('{combi$g1}_{combi$g2}__{combi$label}')) 
```

```{r}
# to organize the final results
anosim_res <- all_anosim %>% 
  bind_rows(.id = 'name') %>% 
  separate(name, into = c('group','pheno'), remove = T, sep = '__') %>% 
  mutate(cohort = case_when(group == 'M_M' ~ "MSK", group == 'P_P' ~ "Upenn", group == 'M_P' ~ "Combined"))

anosim_res %>% 
  write_csv('../data/amplicon/stool/anosim_res.csv')
```

```{r}
# do the pcoa colored by the dominating genus 
g1_ = 'M'  
g2_ = 'P'
label_ = 'domgenus'
df = dat %>% 
  filter(center %in% c(g1_, g2_)) 
cts = df %>% 
  select(sampleid, all_of(keep)) %>% 
  column_to_rownames('sampleid')
phe = df %>% 
  select(matches(label_)) %>% pull(get(label_)) 
phedf = df %>% 
  select(sampleid, matches(label_))

# # plot the pcoa
dist_ <- vegdist(cts, method = 'bray')
bc <- cmdscale(dist_, k = 2)

# the version without the anosim stats
bc %>% 
  as.data.frame() %>% 
  rownames_to_column('sampleid') %>% 
  full_join(phedf, by = 'sampleid') %>% 
  ggplot(aes_string(x = "V1", y = "V2", color =  label_)) + geom_point() +
  labs(title = str_glue('Combined cohort colored by dominating genus')) +
  theme_void() +
  ggsave(str_glue('../figs/beta_{g1_}_{g2_}__{label_}.pdf'))
```


# do a PCOA between the healthy and the CART baseline at asv level

```{r}
library(vdbR)
connect_database('~/dbConfig.txt')
get_table_from_database("healthy_volunteers_ag")
get_table_from_database("asv_alpha_diversity_ag")
stb <- read_csv('../data/amplicon/stool/combined_2_meta.csv')

healthy <- healthy_volunteers_ag %>% 
  inner_join(asv_alpha_diversity_ag, by = c("sampleid", "oligos_id")) 
cts <- get_counts_subset(c(stb$sampleid, healthy %>% pull(sampleid)))

# a total of 75 samples counts there are 3 healthy samples I don't have count ???
nsamp <- cts %>% 
  distinct(sampleid) %>% 
  nrow

all_pheno <- bind_rows(healthy %>% 
  select(sampleid) %>% 
  mutate(grp = 'healthy', center = 'healthy'),
  stb %>% select(sampleid, center)  %>% 
    mutate(grp = 'CART') %>% 
    select(sampleid, grp, center)
  ) %>% 
  ungroup %>% 
  inner_join(asv_alpha_diversity_ag %>% 
               select(path_pool, sampleid))

ttt <- all_pheno %>% 
  ungroup() %>% 
  filter(grp == 'healthy') %>% 
  distinct(sampleid) 

# filter >0.01% in more than 10% samples
keepa <- cts %>% 
  filter(count_relative > 0.0001) %>% 
  count(asv_key) %>% 
  filter(n > floor(nsamp * 0.25)) %>% 
  pull(asv_key)

cts_fil <- cts %>% 
  filter(asv_key %in% keepa) %>% 
  select(sampleid, asv_key,count_relative ) %>% 
  spread(key = 'asv_key', value = 'count_relative', fill = 0) %>% 
  column_to_rownames('sampleid')
library(vegan)
dist_ <- vegdist(cts_fil, method = 'bray')
eigen <- pcoa(dist_)$values$Eigenvalues
percent_var <- signif(eigen/sum(eigen), 3)*100

bc <- cmdscale(dist_, k = 2)

bc %>%
  as.data.frame() %>%
  rownames_to_column('sampleid') %>% 
  ungroup() %>% 
  inner_join(all_pheno) %>% 
  distinct() %>% 
  ggscatter(x = 'V1', y = 'V2', color =  'grp') +
  labs(title = 'PCOA of healthy and CART patients') +
  xlab(paste0("PC 1 [",percent_var[1],"%]")) +
  ylab(paste0("PC 2 [",percent_var[2],"%]")) +
  #theme_void() +
  ggsave('../figs/PCOA(bray-curtis) of healthy and CART patients.pdf')

 
```

```{r}
# a pcoa at asv level for only M and P
cts <- get_counts_subset(c(stb$sampleid))

keepa <- cts %>% 
  filter(count_relative > 0.0001) %>% 
  count(asv_key) %>% 
  filter(n > floor(nsamp * 0.25)) %>% 
  pull(asv_key)

cts_fil <- cts %>% 
  filter(asv_key %in% keepa) %>% 
  select(sampleid, asv_key,count_relative ) %>% 
  spread(key = 'asv_key', value = 'count_relative', fill = 0) %>% 
  column_to_rownames('sampleid')

dist_ <- vegdist(cts_fil, method = 'bray')
eigen <- pcoa(dist_)$values$Eigenvalues
percent_var <- signif(eigen/sum(eigen), 3)*100

bc <- cmdscale(dist_, k = 2)

mp <- bc %>%
  as.data.frame() %>%
  rownames_to_column('sampleid') %>% 
  ungroup() %>% 
  inner_join(all_pheno) %>% 
  distinct(sampleid, .keep_all = T)  %>% 
  mutate(pool = str_extract(path_pool, 'Sample.+/')) %>% 
  mutate(pool = str_replace(pool, 'Sample_','')) %>% 
  mutate(pool = if_else(str_detect(pool, 'IGO'), str_extract(pool, 'IGO.+$'), pool)) %>% 
  mutate(pool = str_replace(pool, '_1/|_comple.+$',''))

mp %>% 
  ggscatter(x = 'V1', y = 'V2', color =  'pool', size = 3) +
  labs(title = 'PCOA of CART patients') +
  xlab(paste0("PC 1 [",percent_var[1],"%]")) +
  ylab(paste0("PC 2 [",percent_var[2],"%]")) +
  #theme_void() +
  ggsave('../figs/PCOA(bray-curtis) (ASV level)of CART patients_pool.pdf', width = 9, height = 9)
```
