==> scripts/49_16s_stacked_barplot.Rmd <==
---
title: "16s taxa stacked bar plot"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
library(tidyverse)
library(vdbR)
connect_database('~/dbConfig.txt')
get_table_from_database("asv_annotation_blast_color_ag");
```


## load the tables and organize the data

```{r}
# my table of the CART stool cohort
stb <- read_csv('../data/amplicon/stool/combined_2_meta.csv')

# get the counts from database and also the color for the asv 
counts_data <- get_counts_subset(stb$sampleid)

counts_data %>% 
  distinct(sampleid)
```


```{r}
dat <- counts_data %>% 
  select(asv_key:count_total, count_relative) %>% 
  left_join(asv_annotation_blast_color_ag %>% 
              select(asv_key,color_label_group_distinct), by = "asv_key")

# there are some ASVs that don't have a color with it, but can use the color for the genus level
color_group <- dat %>% 
  split(is.na(.$color_label_group_distinct))

# find the genus for these asv
get_table_from_database('asv_annotation_blast_ag')

no_color <- color_group %>% 
  pluck('TRUE') %>% 
  distinct(asv_key) %>% 
  inner_join(asv_annotation_blast_ag %>% 
               select(asv_key, genus)) 

# find the colors for these genera
genera_colors <- no_color %>% 
  distinct(genus) %>% 
  inner_join(asv_annotation_blast_color_ag %>% 
               distinct(genus, color_label_group_distinct))


# the full df for the no color genera
no_color_df <- no_color %>% 
  left_join(genera_colors)

no_color_df_full <- color_group %>% 
  pluck('TRUE') %>% 
  select(-color_label_group_distinct) %>% 
  left_join(no_color_df %>% 
              select(- genus))
  
# so if the genus is unknown then it's gonna be assigned "other" gray color  
# the question is do we go one taxa level higher or make a new color base and shades for the new asv
# after discussing with Tsoni, we decided that it's ok to assign gray to the unknown genus 

# merge the new no_color_df_full to the original df
dat <- bind_rows(
  no_color_df_full,
  color_group %>% 
    pluck('FALSE')
)   

dat %>%  write_csv('../data/the_data_to_make_panel_B.csv')
```


```{r}
# the color palette (inherited from Ying, used in lots of project in our lab, the palette used in the NEJM paper Fig 2D https://www.nejm.org/doi/full/10.1056/NEJMoa1900623)

asv_color_set <- asv_annotation_blast_color_ag %>% 
  distinct(color,color_label_group_distinct,color_label_group,color_base) %>% 
  select(color_label_group_distinct, color) %>% 
  deframe()
```

```{r}
# calculate the beta diversity between the samples which deicide the order of the samples in the plot
cbd <- compute_beta_diversity_and_tsne(sampleid = dat$sampleid, 
                                      taxonomy = dat$color_label_group_distinct,
                                      count = dat$count);
#compute beta diversity
cbd$compute_beta_diversity()
```

```{r}
#get beta diversity
d_beta <- cbd$get_betadiversity() 

#compute hierarchical cluster
hc <- hclust(as.dist(d_beta), method = 'complete')

dend <- as.dendrogram(hc)

sample_dendogram_order <- labels(dend)

dat$sampleid = factor(dat$sampleid,levels = sample_dendogram_order)

ggplot(dat,aes(sampleid, count_relative, fill = color_label_group_distinct) ) +
  geom_bar(stat = "identity", position="fill", width = 1) +
  theme_classic() +
  labs(title = 'stacked bar sorted with hclust') +
  theme(axis.text.x = element_text(angle = 90),
        axis.text.y = element_blank(),
        legend.position = "none") +
  scale_fill_manual(values = asv_color_set) +
  ggsave('../figs/amplicon/stacked_bar_sorted_with_hclust.pdf', width = 7, height = 5)

```

```{r}
# the thing that is dividing the samples to lower and higher diversity 
# or just order samples by their diversity
div_order <- stb %>% 
  arrange(simpson_reciprocal) %>% 
  pull(sampleid)

###
# how about splitting the above dendrogram order into the low and higher diversity groups
div_med <- median(stb$simpson_reciprocal)
lower_samp <- stb %>% 
  filter(simpson_reciprocal <= div_med) %>% 
  pull(sampleid)
lower_samp_o <- sample_dendogram_order[sample_dendogram_order %in% lower_samp]
higher_samp_o <- sample_dendogram_order[!sample_dendogram_order %in% lower_samp]

dat$sampleid = factor(dat$sampleid,levels = c(lower_samp_o, higher_samp_o))

ggplot(dat,aes(sampleid, count_relative, fill = color_label_group_distinct) ) +
  geom_bar(stat = "identity", position="fill", width = 1) +
  theme_classic() +
  labs(title = 'stacked bar sorted with hclust and divided to lower and higher diversity') +
  theme(axis.text.x = element_text(angle = 90),
        axis.text.y = element_blank(),
        legend.position = "none") +
  scale_fill_manual(values = asv_color_set) +
  ggsave('../figs/amplicon/stacked_bar_sorted_with_hclust_lower_and_higher_diversity.pdf', width = 7, height = 5)
```
