==> scripts/54_rethinking_binary.Rmd <==
---
title: "the rethinking way of doing the binary outcome modeling"
author: "Anqi Dai"
date: "`r Sys.Date()`"
output:
  html_document:
    toc: true
    toc_float: true
    toc_depth: 3 
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, message = F, warning = F)
```

```{r}
library(rethinking)
library(tidyverse) 
library(ggpubr)
options(mc.cores = parallel::detectCores())
rstan_options(auto_write = TRUE)
library(tidybayes)
```

# CR/Toxicity ~ alpha diversity + center

## Toxicity 

```{r}
meta <- read_csv('../data/amplicon/stool/combined_2_meta.csv') %>% 
  mutate(logdiv_s = as.numeric(scale(log(simpson_reciprocal)))) %>%  
  mutate(tox = if_else(toxicity == 'yes', 1, 0),
         cr100 = if_else(cr_d100 == 'yes', 1, 0),
         loca = if_else(center == 'M', 1, 2)) %>%  # MSK 1 ; Upenn 2
  mutate(center = factor(center),
         toxicity = factor(toxicity, levels = c('no','yes')),
         cr_d100 = factor(cr_d100, levels = c('no','yes'))) 

meta %>% write_csv('../data/amplicon/stool/combined_2_meta_expanded.csv')
```

```{r}
# visualize the var distribution 
table(meta$cr_d100)  
table(meta$toxicity)
#hist(meta$simpson_reciprocal)
#hist(log(meta$simpson_reciprocal)) # I think I should use log transformed
table(meta$center)
table(meta$toxicity, meta$center)
table(meta$cr_d100, meta$center)   
range(meta$logdiv_s)
table(meta$crs3)
table(meta$crsgrade, meta$center) 
```

```{r}
set.seed(123)
dat_list <- list(
    tox = meta$tox,
    crs3 = meta$crs3,
    cr100 = meta$cr100,
    location = meta$loca,
    logdiv_s = meta$logdiv_s
    )

mtox <- ulam(
    alist(
        tox ~ dbinom( 1 , p ) ,
        logit(p) <- b*logdiv_s + a[location] ,
        b ~ dnorm( 0 , 2), 
        a[location] ~ dnorm( 0 , 0.5 )
    ) , data=dat_list , chains=4 , log_lik=TRUE , cores = 16)

# the std of b is 2, since we assume the microbiome could have big impact on immunology (based on our previous understanding and knowledge )
precis(mtox,  prob = 0.95)
plot(precis(mtox, depth = 2, prob = 0.95 ) , pars = ('b'), labels = c('simpson_reciprocal'))

# make this plot myself with a pre-defined range in the axis
df <- precis(mtox,  prob = 0.95)  %>% as.tibble()
df <- df %>%
  mutate(variable = 'simpson_reciprocal')

#remove row number 1 (The intercept)
df %>% 
  ggplot(aes(x = reorder(variable, mean), y = mean)) +
  geom_point(shape = 15,
             size  = 4, width = 0.1,
             position = "dodge", color="black") +
  geom_errorbar(aes(ymin  = `2.5%`,
                    ymax  = `97.5%`),
                width = 0.1,
                size  = 0.7,
                position = "dodge", color="turquoise4") +
  theme(axis.title = element_text(face = "bold")) +
  xlab("Variables") + ylab("Coefficient with 95% CI") +
  coord_flip(ylim = c(-0.7, 1)) +
  #geom_hline(yintercept = 1, color = "red", size = 1) +
   theme(axis.title = element_text(size = 0)) +
  theme(axis.text = element_text(size = 14)) +
  labs(title = 'Toxicity') 

ggsave('../figs/amplicon/bayesian_div_tox.pdf', width = 5, height = 4)
```
There is baseline difference between penn and Msk and MSK has higher tox rage in general.
it doesn't seem diversity has a clear association with tox rate after correcting for center wise difference


### Prior

```{r}
# doing a prior check
prior <- extract.prior( mtox , n=1e4 ) 
# the diff in the probabilty of yes to toxicity when there are two std diff in diversity in each of the centers 
# for msk  
div_value <- c(-1, 1)

# make it into a function so that I can do the m and p the same time
prior_tox <- list(m = 1, p = 2) %>% 
  map(function(center_){
    cols = seq(1, length(div_value)) %>% 
      set_names(seq(1, length(div_value))) %>% 
      map(function(idx){
         inv_logit( prior$b * div_value[idx] + prior$a[, center_] ) 
      })
    
    diff =  cols[[1]] -cols[[2]] 
    ret = quantile(diff , probs = c(0.025, 0.5, 0.975))
    return(diff)
  }) 
boxplot(prior_tox$p)
hist(c(prior_tox$m$lowdiv, prior_tox$p$lowdiv)) 
hist(c(prior_tox$m$highdiv, prior_tox$p$highdiv))
```


### Posterior

```{r}
# posterior prediction check
post <- extract.samples(mtox)
post_tox <- list(m = 1, p = 2) %>% 
  map(function(center_){
    cols = seq(1, length(div_value)) %>% 
      set_names(seq(1, length(div_value))) %>% 
      map(function(idx){
         inv_logit( post$b * div_value[idx] + post$a[, center_] ) 
      })
    
    diff = abs( cols[[1]] -cols[[2]] )
    ret = quantile(diff , probs = c(0.025, 0.5, 0.975))
    return(cols)
  }) # here I'm returning the cols instead of the diff 
```

```{r}
# posterior check for tox
# see the coeff distribution directly without going to the difference 
c('m','p') %>% 
  map(function(x){ tibble(
           post_low = post_tox %>% pluck(x) %>% pluck(1),
           post_high = post_tox %>% pluck(x) %>% pluck(2)
          )  %>% gather() %>% 
      gghistogram(x = 'value',bins = 30, fill = 'key', palette = 'nejm', color = 'white',
                  xlab = 'Predicted probability of toxicity', ylab = 'Probability density',
                  title = str_glue('{x}')) +
      scale_x_continuous(breaks = seq(0, 1, 0.1)) })
```


```{r}
# combine M and P since they look very similar and also save some space
low_M <- post_tox %>% pluck('m') %>% pluck(1)
low_P <- post_tox %>% pluck('p') %>% pluck(1)
high_M <- post_tox %>% pluck('m') %>% pluck(2)
high_P <- post_tox %>% pluck('p') %>% pluck(2)
low_all <- bind_rows(
  tibble(value = low_M),
  tibble(value = low_P)
) %>% mutate(grp = 'low')

high_all <- bind_rows(
  tibble(value = high_M),
  tibble(value = high_P)
)%>% mutate(grp = 'high')

tox_all <- bind_rows(
  low_all,
  high_all
)

tox_all  %>% 
  gghistogram(x = 'value',bins = 30, fill = 'grp', palette = 'nejm', color = 'white',
                  xlab = 'Predicted probability of toxicity', ylab = 'Probability density',
                  title = '') +
      scale_x_continuous(breaks = seq(0, 1, 0.1)) +
  ggsave('../figs/bayesican_tox_combined.pdf', width = 6, height = 4)
```


```{r}
# visualize the prior and post together
# tox <- bind_rows(
#   bind_cols(prior_tox, grp ='prior', quant = c(0.025, 0.5, 0.975)),
#   bind_cols(post_tox, grp ='post', quant = c(0.025, 0.5, 0.975))
# )  %>% 
#   gather('center', 'value', m:p) %>% 
#   spread('quant', 'value') %>% 
#   mutate(grp = factor(grp, levels = c('prior','post')))
# tox
```
```{r}
# tox %>% 
#   ggscatter(x = 'center', y = '0.5', color = 'red', facet.by = 'grp', ylab = '', title = 'Difference in probability of responding yes when the diversity is 2 std apart') +
#   geom_pointrange(aes( ymin  = `0.025`,ymax  = `0.975`), fatten = 5, col = 'red') +
#   ylim(0,1) 
```


```{r}
# the raw proportion of yes toxicity in each place
# rawtox <- meta  %>% 
#   count(center, toxicity) %>% 
#   left_join(meta %>% 
#               count(center, name = 'total'), by = 'center') %>% 
#   group_by(center, toxicity) %>% 
#   summarise(frac = round(n/total, 2)) %>% 
#   filter(toxicity == 'yes')
  

```


## CR_d100      

```{r}
mcr <- ulam( 
    alist( 
        cr100 ~ dbinom( 1 , p ) ,
        logit(p) <- b*logdiv_s + a[location] ,
        b ~ dnorm( 0 , 2),  
        a[location] ~ dnorm( 0 , 0.5 )
    ) , data=dat_list , chains=4 , log_lik=TRUE , cores = 16)

precis(mcr,  prob = 0.95)
plot(precis(mcr, depth = 2, prob = 0.95 ) , pars = ('b'), labels = c('simpson_reciprocal'))
# there isn't much difference in baseline level between the centers
# diversity seems to be positively associating with CR when the prior is too large sd
# if using appropriate prior the association is less sure 
 
```

```{r}
# make this plot myself with a pre-defined range in the axis
df <- precis(mcr,  prob = 0.95)  %>% as.tibble()
df <- df %>%
  mutate(variable = 'simpson_reciprocal')

#remove row number 1 (The intercept)
df %>% 
  ggplot(aes(x = reorder(variable, mean), y = mean)) +
  geom_point(shape = 15,
             size  = 4, width = 0.1,
             position = "dodge", color="black") +
  geom_errorbar(aes(ymin  = `2.5%`,
                    ymax  = `97.5%`),
                width = 0.1,
                size  = 0.7,
                position = "dodge", color="turquoise4") +
  theme(axis.title = element_text(face = "bold")) +
  xlab("Variables") + ylab("Coefficient with 95% CI") +
  coord_flip(ylim = c(-0.7, 1)) +
  #geom_hline(yintercept = 1, color = "red", size = 1) +
   theme(axis.title = element_text(size = 0)) +
  theme(axis.text = element_text(size = 14)) +
  labs(title = 'CR') 
ggsave('../figs/amplicon/bayesian_div_cr.pdf', width = 5, height = 4)
```


```{r}
# doing a prior check
prior <- extract.prior( mcr , n=1e4, cores =16 ) 

prior_cr <- list(m = 1, p = 2) %>% 
  map(function(center_){
    cols = seq(1, length(div_value)) %>% 
      set_names(seq(1, length(div_value))) %>% 
      map(function(idx){
         inv_logit( prior$b * div_value[idx] + prior$a[, center_] ) 
      })
    
    diff = cols[[1]] -cols[[2]] 
    ret = quantile(diff , probs = c(0.025, 0.5, 0.975))
    return(diff)
  })


boxplot(prior_cr$p)
hist(c(prior_cr$m$lowdiv, prior_cr$p$lowdiv)) 
hist(c(prior_cr$m$highdiv, prior_cr$p$highdiv)) 
```
### posterior 

```{r}
# posterior prediction check
post <- extract.samples(mcr)
post_cr <- list(m = 1, p = 2) %>% 
  map(function(center_){
    cols = seq(1, length(div_value)) %>% 
      set_names(seq(1, length(div_value))) %>% 
      map(function(idx){
         inv_logit( post$b * div_value[idx] + post$a[, center_] ) 
      }) 
    diff =  cols[[1]] -cols[[2]] 
    return(diff)
  }) 

hist(c(post_cr$m, post_cr$p)) 
hist(c(post_cr$m$highdiv, post_cr$p$highdiv))
```


```{r}
# if putting them together (the prior and the post)
both <- bind_rows(
  data_frame(
      coeff = c(prior_cr$m, prior_cr$p),
      grp = 'prior'
    ),
  data_frame(
      coeff = c(post_cr$m, post_cr$p),
      grp = 'post'
    )
)
 
both %>% 
    ggplot(aes(x = coeff, y = grp, color = grp)) +
    stat_pointinterval(.width = c(.66, .95))
```


```{r}
# plot the histogram of the probs corresponding to low and high diversity
# msk
tibble(
  prior_low = prior_cr %>% pluck('m') %>% pluck(1),
  prior_high = prior_cr %>% pluck('m') %>% pluck(2)
)%>% gather() %>% 
  gghistogram(x = 'value')+
  facet_grid(key ~ .)

# posterior 
# do the M and P in one function
c('m','p') %>% 
  map(function(x){ tibble(
           post_low = post_cr %>% pluck(x) %>% pluck(1),
           post_high = post_cr %>% pluck(x) %>% pluck(2)
          )  %>% gather() %>% 
      gghistogram(x = 'value',bins = 30, fill = 'key', palette = 'nejm', color = 'white',
                  xlab = 'Predicted probability of CR d100', ylab = 'Probability density',
                  title = str_glue('{x}')) +
      scale_x_continuous(breaks = seq(0, 1, 0.1)) })
```

```{r}
# combine M and P
low_all <- tibble(
  value = c(post_cr %>% pluck('m') %>% pluck(1), post_cr %>% pluck('p') %>% pluck(1))
) %>% mutate(grp = 'low')

high_all <- tibble(
  value = c(post_cr %>% pluck('m') %>% pluck(2), post_cr %>% pluck('p') %>% pluck(2))
) %>% mutate(grp = 'high')

cr_all <- bind_rows(
  low_all,
  high_all
)


cr_all  %>% 
  gghistogram(x = 'value',bins = 30, fill = 'grp', palette = 'nejm', color = 'white',
                  xlab = 'Predicted probability of CR d100', ylab = 'Probability density',
                  title = '') +
      scale_x_continuous(breaks = seq(0, 1, 0.1)) +
  ggsave('../figs/bayesican_cr_combined.pdf', width = 6, height = 4)
```


```{r}
# the difference of the probability 
post_cr <- list(m = 1, p = 2) %>% 
  map(function(center_){
    cols = seq(1, length(div_value)) %>% 
      set_names(seq(1, length(div_value))) %>% 
      map(function(idx){
         inv_logit( post$b * div_value[idx] + post$a[, center_] ) 
      })
    
    diff = abs( cols[[1]] -cols[[2]] )
    ret = quantile(diff , probs = c(0.025, 0.5, 0.975))
    return(ret)
  })
```

 
```{r}
#the below is when the returned value is the difference between the low and high
diff_prob_post <- bind_rows(
  bind_cols(prior_cr, grp ='prior', quant = c(0.025, 0.5, 0.975)),
  bind_cols(post_cr, grp ='post', quant = c(0.025, 0.5, 0.975))
)  %>%
  gather('center', 'value', m:p) %>%
  spread('quant', 'value') %>%
  mutate(grp = factor(grp, levels = c('prior','post'))) %>%
  # only showing the post check value
  filter(grp == 'post')
# mean at both centers are 0.20 actually 

diff_prob_post  %>% 
  ggscatter(x = 'center', y = '0.5', color = 'red', facet.by = 'grp', ylab = '', title = 'Difference in probability of responding yes when the diversity is 2 sd apart') +
  geom_pointrange(aes( ymin  = `0.025`,ymax  = `0.975`), fatten = 5, col = 'red') 
```


```{r}
#  The two patients with highest and lowest diversity
# post only 
# lowest appear in one patient in p; highest is in one patient in M  
div_l <- min(meta$logdiv_s)
pl <- inv_logit( post$b * div_l + post$a[, 2] ) 

div_h <- max(meta$logdiv_s) 
ph <- inv_logit( post$b * div_h + post$a[, 1] ) 

bind_rows(
  tibble(value = pl, grp = 'lowest'),
  tibble(value = ph, grp = 'highest')
) %>% 
  gghistogram(x = 'value',bins = 30, fill = 'grp', palette = 'nejm', color = 'white', add = 'mean', xlab = 'Predicted probability of CR d100', ylab = 'Probability density',
              title = 'Predicted probability of CR d100 between the pt with highest div and the pt with lowest') 

mean(pl)
mean(ph)
```
