==> scripts/55_rethinking_abx.Rmd <==
---
title: "rethinking antibiotics"
author: "Anqi Dai"
date: "`r Sys.Date()`"
output:
  html_document:
    toc: true
    toc_float: true
    toc_depth: 3
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, warning = F, message = F)
```
 
```{r}
library(rethinking)
library(tidyverse)  
library(ggpubr)
```

# CR/Toxicity ~ abx + center

## Toxicity

```{r}
# none of the patients from Penn received the 3 anaerobe-targeting antibiotics in the 4 weeks prior to CAR T cell infusion
# msk table
m <- readxl::read_excel('/Volumes/castoricenter/CAR-Tcell_and_Microbiota/MSK/antibiotics/sean_table/abx_table_msk_sean.xlsx') %>% 
  select(pt_id, anaerobe_targeting)

meta <- read_csv('../data/amplicon/stool/combined_2_meta_expanded.csv') %>% 
  left_join(m) %>% 
  mutate(anaerobe_targeting = if_else(is.na(anaerobe_targeting), 'no', anaerobe_targeting)) %>% 
  mutate(anaerobe = if_else(anaerobe_targeting == 'no', 0, 1)) 

set.seed(123)
dat_list <- list(
    tox = meta$tox,
    cr100 = meta$cr100,
    location = meta$loca,
    anaerobe = meta$anaerobe
    )

meta %>% 
  count(center, anaerobe_targeting) 
```

```{r}   
atox <- ulam(
    alist(
        tox ~ dbinom( 1 , p ) ,
        logit(p) <- b*anaerobe + a[location] ,
        b ~ dnorm( 0 , 1.5), 
        a[location] ~ dnorm( 0 , 0.5 )
    ) , data=dat_list , chains=4 , log_lik=TRUE , cores = 8)
precis(atox,prob = 0.95)
plot(precis(atox, depth = 2 ,prob = 0.95) ) 
# seems anaerobe abx is positively linked with toxicity 
```
### Posterior check

```{r}
abx_val <- c(0, 1)
post <- extract.samples(atox)
post_tox <- list(m = 1, p = 2) %>% 
  map(function(center_){
    cols = seq(1, length(abx_val)) %>% 
      set_names(seq(1, length(abx_val))) %>% 
      map(function(idx){
         inv_logit( post$b * abx_val[idx] + post$a[, center_] ) 
      })
    
    #diff = abs( cols[[1]] -cols[[2]] )
    #ret = quantile(diff , probs = c(0.025, 0.5, 0.975))
    return(cols)
  })

c('m','p') %>% 
  map(function(x){ tibble(
           post_no = post_tox %>% pluck(x) %>% pluck(1),
           post_yes = post_tox %>% pluck(x) %>% pluck(2)
          )  %>% gather() %>% 
      gghistogram(x = 'value',bins = 30, fill = 'key', palette = 'nejm', color = 'white',xlab = 'Predicted probability of toxicity', ylab = 'Probability density', title = str_glue('{x}')) +
      scale_x_continuous(breaks = seq(0, 1, 0.1)) })

# to find out the mean for MSK
df <- tibble(
           post_no = post_tox %>% pluck('m') %>% pluck(1),
           post_yes = post_tox %>% pluck('m') %>% pluck(2)
          )  %>% gather()  %>% 
  group_by(key) %>% 
  summarise(meancoeff = mean(value))
```

## CR day 100 

```{r}
acr <- ulam(  
    alist(
        cr100 ~ dbinom( 1 , p ) ,
        logit(p) <- b*anaerobe + a[location] ,
        b ~ dnorm( 0 , 1.5), 
        a[location] ~ dnorm( 0 , 0.5 )
    ) , data=dat_list , chains=4 , log_lik=TRUE , cores = 8)
precis(acr,  prob = 0.95)
plot(precis(acr, depth = 2, prob = 0.95) ) 
# nothing interesting 
```
### Posterior check

```{r}
abx_val <- c(0, 1)
post <- extract.samples(acr)
post_cr <- list(m = 1, p = 2) %>% 
  map(function(center_){
    cols = seq(1, length(abx_val)) %>% 
      set_names(seq(1, length(abx_val))) %>% 
      map(function(idx){
         inv_logit( post$b * abx_val[idx] + post$a[, center_] ) 
      })
    
    #diff = abs( cols[[1]] -cols[[2]] )
    #ret = quantile(diff , probs = c(0.025, 0.5, 0.975))
    return(cols)
  })

c('m','p') %>% 
  map(function(x){ tibble(
           post_no = post_cr %>% pluck(x) %>% pluck(1),
           post_yes = post_cr %>% pluck(x) %>% pluck(2)
          )  %>% gather() %>% 
      gghistogram(x = 'value',bins = 30, fill = 'key', palette = 'nejm', color = 'white',
                  xlab = 'Predicted probability of CR d100', ylab = 'Probability density', title = str_glue('{x}')) +
      scale_x_continuous(breaks = seq(0, 1, 0.1))})

df <- tibble(
           post_no = post_cr %>% pluck('m') %>% pluck(1),
           post_yes = post_cr %>% pluck('m') %>% pluck(2)
          )  %>% gather()  %>% 
  group_by(key) %>% 
  summarise(meancoeff = mean(value))
```
