==> scripts/56_rethinking_genera.Rmd <==
---
title: "rethinking and the 5 genera"
author: "Anqi Dai"
date: "`r Sys.Date()`"
output:
  html_document:
    toc: true
    toc_float: true
    toc_depth: 3
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, warning = F, message = F)
```

```{r}
library(rethinking)
library(tidyverse)  
library(ggpubr)
set.seed(123)
```

# CR/Toxicity ~ Akkermansia + Bacteroides + Enterococcus + Faecalibacterium + Ruminococcus + center

## Toxicity

```{r}
genera <- read_csv('../data/amplicon/stool/combined_5_genera.csv')

# do not standardize the log transformed relative abundance 
meta <- read_csv('../data/amplicon/stool/combined_2_meta_expanded.csv')  %>% 
  inner_join(genera)

# wanna see the range of the 5 genera log 10 relab
meta %>% select(Akkermansia:Ruminococcus) %>% 
  summary
```

```{r}
set.seed(123)   
dat_list <- list(
    tox = meta$tox,
    cr100 = meta$cr100,
    location = meta$loca,
    Akkermansia = meta$Akkermansia,
    Bacteroides = meta$Bacteroides,
    Enterococcus = meta$Enterococcus,
    Faecalibacterium = meta$Faecalibacterium,
    Ruminococcus = meta$Ruminococcus
    ) 
```

```{r}
gtox <- ulam(  
    alist(
        tox ~ dbinom( 1 , p ) ,
        logit(p) <- ba*Akkermansia + bb*Bacteroides + be*Enterococcus + bf*Faecalibacterium + br*Ruminococcus + a[location] ,
        ba ~ dnorm( 0 , 1),
        bb ~ dnorm( 0 , 1),
        be ~ dnorm( 0 , 1),
        bf ~ dnorm( 0 , 1),
        br ~ dnorm( 0 , 1),
        a[location] ~ dnorm( 0 , 0.5 )
    ) , data=dat_list , chains=4 , log_lik=TRUE , cores = 16)
precis(gtox, prob = 0.95)   
plot(precis(gtox, depth = 2, prob = 0.95) , pars = c('ba','bb','be','bf','br'), labels = c('Akkermansia', 'Bacteroides','Enterococcus','Faecalibacterium','Ruminococcus'))  
# indicative of the Bacteroides link with toxicity  
```
   
```{r}
# make this plot myself with a pre-defined range in the axis
df <- precis(gtox,  prob = 0.95)  %>% as.tibble()
df <- df %>%
  mutate(variable = c('Akkermansia', 'Bacteroides','Enterococcus','Faecalibacterium','Ruminococcus'))

#remove row number 1 (The intercept)
df %>% 
  ggplot(aes(x = variable, y = mean)) +
  geom_point(shape = 15,
             size  = 4, width = 0.1,
             position = "dodge", color="black") +
  geom_errorbar(aes(ymin  = `2.5%`,
                    ymax  = `97.5%`),
                width = 0.1,
                size  = 0.7,
                position = "dodge", color="turquoise4") +
  theme(axis.title = element_text(face = "bold")) +
  xlab("Variables") + ylab("Coefficient with 95% CI") +
  coord_flip(ylim = c(-1.5, 1.5)) +
  geom_hline(yintercept = 0, color = "red", size = 1) +
   theme(axis.title = element_text(size = 0)) +
  theme(axis.text = element_text(size = 14)) +
  labs(title = 'Toxicity') 
ggsave('../figs/amplicon/bayesian_genera_tox_log10.pdf', width = 5, height = 4)
```

### Prior check

I will just do something similar to the nutrition project : the prior annd post coeff forest plot

```{r}
theme_set(theme_tidybayes() + cowplot::panel_border())
# logit(p) <- ba*Akkermansia + bb*Bacteroides + be*Enterococcus + bf*Faecalibacterium + br*Ruminococcus + a[location]
prior <- extract.prior( gtox , n=4000 ) %>% as.tibble()  %>% select(-a) %>% mutate(grp = 'prior')
post <- extract.samples(gtox) %>% as.tibble()  %>% select(-a) %>% mutate(grp = 'post')
# comparing prior and post 
tox <- bind_rows(prior, post) %>% 
  gather('term', 'coeff', ba:br) %>% 
  mutate(genus = case_when(
    term == 'ba' ~ 'Akkermansia',
    term == 'bb' ~ 'Bacteroides',
    term == 'be' ~ 'Enterococcus',
    term == 'bf' ~ 'Faecalibacterium',
    term == 'br' ~ 'Ruminococcus'
  ))

post_order <- tox %>% 
  filter(grp == 'post') %>% 
  group_by(term, genus) %>% 
  summarise(q50 = median(coeff)) %>% 
  arrange(-q50) %>% 
  pull(genus)

tox %>%
  mutate(genus = factor(genus, levels = post_order)) %>% 
  ggplot(aes(x = coeff, y = grp, color = grp)) +
  stat_pointinterval(.width = c(.66, .95)) +
  scale_color_manual(values = c('#EC0000','#00468B')) +
  geom_vline(xintercept = 0, col = 'gray', linetype = 'dashed') +
  facet_grid(genus ~ .) +
  labs(x = 'Regression coefficients',
       y = '') +
  theme(legend.position = 'none')
```


### posterior check 
   
```{r}

meta_ <- meta %>% 
  select(Akkermansia:Ruminococcus)

# the post coeffs in df format
postdf <- bind_cols(ba = post$ba, bb = post$bb, be = post$be, bf = post$bf, br = post$br, am = post$a[,1])

N <- 100
# for the top and bottom quantile range of Bacteroides relab
ba_top <- meta_ %>% 
  filter(Bacteroides >= quantile(meta$Bacteroides, 0.9))
ba_bot <- meta_ %>% 
  filter(Bacteroides <= quantile(meta$Bacteroides, 0.1))

per_pt_sample_post_prob <- function(lga_, lgb_, lge_, lgf_ , lgr_){
  # the input is the log relab of the 5 genera for that patient
  post_samp <- postdf %>% 
    sample_n(size = N, replace = F)
  
  ret = pmap(post_samp, function(ba, bb, be, bf, br, am) {
    inv_logit( ba * lga_ + bb * lgb_ + be * lge_ +  bf * lgf_ + br * lgr_ + am ) 
  }) %>% 
    set_names(seq(1, N)) %>% 
    bind_rows() %>% 
    gather()
  return(ret)
}

# for the top quantile patients 
res_top <- pmap(ba_top, function(Akkermansia, Bacteroides, Enterococcus, Faecalibacterium, Ruminococcus){
  per_pt_sample_post_prob(Akkermansia, Bacteroides, Enterococcus, Faecalibacterium, Ruminococcus)
}) %>% 
  set_names(paste('P', seq(1, nrow(ba_top)), sep = '')) %>% 
  bind_rows(.id = 'pt')

# for the bottom quantile patients  
res_bot <- pmap(ba_bot, function(Akkermansia, Bacteroides, Enterococcus, Faecalibacterium, Ruminococcus){
  per_pt_sample_post_prob(Akkermansia, Bacteroides, Enterococcus, Faecalibacterium, Ruminococcus)
}) %>% 
  set_names(paste('P', seq(1, nrow(ba_bot)), sep = '')) %>% 
  bind_rows(.id = 'pt')   

# plot them together in one 
bacteroides_top_bottom_10 <- bind_rows(
  res_top %>% mutate(grp = 'top'),
  res_bot %>% mutate(grp = 'bot')
) %>% 
  gghistogram(x = 'value',bins = 30, fill = 'grp', palette = 'nejm', color = 'white', add = 'mean',
                  xlab = 'Predicted probability of Toxicity', ylab = 'Probability density') 

ggsave('../figs/predicted_tox_bacteroides_top_bottom_10.pdf', width = 7, height = 5)

bind_rows(
  res_top %>% mutate(grp = 'top'),
  res_bot %>% mutate(grp = 'bot')
) %>% 
  group_by(grp) %>% 
  summarise(ave = mean(value))


```



## CR

```{r}
gcr <- ulam(
    alist(
        cr100 ~ dbinom( 1 , p ) ,
        logit(p) <- ba*Akkermansia + bb*Bacteroides + be*Enterococcus + bf*Faecalibacterium + br*Ruminococcus + a[location] ,
        ba ~ dnorm( 0 , 1),
        bb ~ dnorm( 0 , 1),
        be ~ dnorm( 0 , 1),
        bf ~ dnorm( 0 , 1),
        br ~ dnorm( 0 , 1),
        a[location] ~ dnorm( 0 , 0.5 )
    ) , data=dat_list , chains=4 , log_lik=TRUE , cores = 16)
precis(gcr, prob = 0.95) 
plot(precis(gcr, depth = 2, prob = 0.95) ) 
# I think the Ruminococcus still holds for the CR
```

```{r}
# make this plot myself with a pre-defined range in the axis
df <- precis(gcr,  prob = 0.95)  %>% as.tibble()
df <- df %>%
  mutate(variable = c('Akkermansia', 'Bacteroides','Enterococcus','Faecalibacterium','Ruminococcus'))

#remove row number 1 (The intercept)
df %>% 
  ggplot(aes(x = variable, y = mean)) +
  geom_point(shape = 15,
             size  = 4, width = 0.1,
             position = "dodge", color="black") +
  geom_errorbar(aes(ymin  = `2.5%`,
                    ymax  = `97.5%`),
                width = 0.1,
                size  = 0.7,
                position = "dodge", color="turquoise4") +
  theme(axis.title = element_text(face = "bold")) +
  xlab("Variables") + ylab("Coefficient with 95% CI") +
  coord_flip(ylim = c(-1.5, 1.5)) +
  geom_hline(yintercept = 0, color = "red", size = 1) +
   theme(axis.title = element_text(size = 0)) +
  theme(axis.text = element_text(size = 14)) +
  labs(title = 'CR') 
ggsave('../figs/amplicon/bayesian_genera_cr_log10.pdf', width = 5, height = 4)
```

### posterior check 

```{r}
# ruminococcus relab quantile
meta_ <- meta %>% 
  select(Akkermansia:Ruminococcus)

# the post coeffs in df format
postdf <- bind_cols(ba = post$ba, bb = post$bb, be = post$be, bf = post$bf, br = post$br, am = post$a[,1])

# to responde to Sean's comments to not use max and min "extreme"patient data
# use the 90% and 10% quantile patient data
# Jonas: how do you know that the patient falling onto the 90%tile is not “exceptional” and therefore an “outlier” in another feature that your model includes, besides ruminococcus, he suggests to sample in that 0-10% quantile range patient data for example
# It is 100 jointly sampled coefficients multiplied with the composition of bacteria in the current patient
N <- 100
# for the top and bottom quantile range of ruminococcus relab
ru_top <- meta_ %>% 
  filter(Ruminococcus >= quantile(meta$Ruminococcus, 0.9))
ru_bot <- meta_ %>% 
  filter(Ruminococcus <= quantile(meta$Ruminococcus, 0.1))

# sample in the top and bot patients 100 times
ru_top_samp <- ru_top %>% 
    sample_n(size = N, replace = T)
ru_bot_samp <- ru_bot %>% 
    sample_n(size = N, replace = T)

# a function for every patient data generate 100 post prob of CR (sampling from the post 100 times)

per_pt_sample_post_prob <- function(lga_, lgb_, lge_, lgf_ , lgr_){
  # the input is the log relab of the 5 genera for that patient
  post_samp <- postdf %>% 
    sample_n(size = N, replace = F)
  
  ret = pmap(post_samp, function(ba, bb, be, bf, br, am) {
    inv_logit( ba * lga_ + bb * lgb_ + be * lge_ +  bf * lgf_ + br * lgr_ + am ) 
  }) %>% 
    set_names(seq(1, N)) %>% 
    bind_rows() %>% 
    gather()
  return(ret)
}

# for the top quantile patients 
res_top <- pmap(ru_top, function(Akkermansia, Bacteroides, Enterococcus, Faecalibacterium, Ruminococcus){
  per_pt_sample_post_prob(Akkermansia, Bacteroides, Enterococcus, Faecalibacterium, Ruminococcus)
}) %>% 
  set_names(paste('P', seq(1, nrow(ru_top)), sep = '')) %>% 
  bind_rows(.id = 'pt')

# for the bottom quantile patients
res_bot <- pmap(ru_bot, function(Akkermansia, Bacteroides, Enterococcus, Faecalibacterium, Ruminococcus){
  per_pt_sample_post_prob(Akkermansia, Bacteroides, Enterococcus, Faecalibacterium, Ruminococcus)
}) %>% 
  set_names(paste('P', seq(1, nrow(ru_bot)), sep = '')) %>% 
  bind_rows(.id = 'pt')
```

```{r}
# plotting and visualizing
# visualizing each patient
res_top %>% 
  gghistogram(x = 'value', fill = 'pt', add_density = T,
              title = '100 sampled patients that had top 10% Ruminococcus relab',
              xlab = 'Probability of having CR d100', ylab = 'Probability density') +
  theme(legend.position = 'none')

 
# averaging over all the patients
res_top %>% 
  gghistogram(x = 'value', 
              #add_density = T,
              title = '100 sampled patients that had top 10% Ruminococcus relab',
              xlab = 'Probability of having CR d100', ylab = 'Probability density') +
  theme(legend.position = 'none')

mean(res_top$value)
```


```{r}
# visualizing each patient
res_bot %>% 
  gghistogram(x = 'value', fill = 'pt', add_density = T,
              title = '100 sampled patients that had bottom 10% Ruminococcus relab',
              xlab = 'Probability of having CR d100', ylab = 'Probability density') +
  theme(legend.position = 'none')

# averaging over all the patients
res_bot %>% 
  gghistogram(x = 'value', 
              #add_density = T,
              title = '100 sampled patients that had bottom 10% Ruminococcus relab',
              xlab = 'Probability of having CR d100', ylab = 'Probability density') +
  theme(legend.position = 'none')

mean(res_bot$value)
```

```{r}
# plot them together in one 
bind_rows(
  res_top %>% mutate(grp = 'top'),
  res_bot %>% mutate(grp = 'bot')
) %>% 
  gghistogram(x = 'value',bins = 30, fill = 'grp', palette = 'nejm', color = 'white', add = 'mean',
                  xlab = 'Predicted probability of CR d100', ylab = 'Probability density')
  ggsave('../figs/predicted_CR_Ruminococcus_top_bottom_10.pdf', width = 7, height = 5)

bind_rows(
  res_top %>% mutate(grp = 'top'),
  res_bot %>% mutate(grp = 'bot')
)  %>% 
  group_by(grp) %>% 
  summarise(ave = mean(value))
```
